### 2024-07-02 #6913 @danielskatz
@freifrauvonbleifrei - I'll go ahead and edit this one. Please suggest 4-5 potential reviewers, perhaps from [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission

If you have suggestions where you know a person's github handle, please mention it here in this thread (without tagging them with an @). If not, name and affiliation are ok too.
