### 2024-07-12 #6913 @danielskatz
👋 @dokempf & @EmilyBourne - would one or both of you be willing to review this JOSS submission?
