### 2024-07-12 #6913 @danielskatz
👋 @josemanuel22 & @jakelangham - would one or both of you be willing to review this JOSS submission?
