### 2024-07-22 #7018 @editorialbot
[REVIEW]: DisCoTec: Distributed higher-dimensional HPC simulations with the sparse grid combination technique
**Submitting author:** @freifrauvonbleifrei (<a href="http://orcid.org/0000-0002-0186-4340">Theresa Pollinger</a>)
**Repository:** https://github.com/SGpp/DisCoTec/
**Branch with paper.md** (empty if default branch): feature_joss_submission
**Version:** 1.1.0
