### 2024-08-07 #7018 @danielskatz
👋 @jakelangham - note that you need to use the command `@editorialbot generate my checklist` to create your review checklist, in order to get started. @editorialbot commands need to be the first thing in a new comment.
