### 2024-08-08 935ae62
CMake: add target / arch flag option DISCOTEC_COMPILE_ARCH (may be empty); -march=native is still default
