### 2024-08-19 SGpp/DisCoTec#127 @EmilyBourne
[JOSS] Community guidelines - add a CONTRIBUTING.md file
Regarding  https://github.com/openjournals/joss-reviews/issues/7018 - "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I recommend adding a [CONTRIBUTING.md](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file to the repository to outline how others might contribute to this repository.
