### 2024-08-19 SGpp/DisCoTec#128 @EmilyBourne
[JOSS] Contributing and authorship
Regarding https://github.com/openjournals/joss-reviews/issues/7018 - "Contribution and authorship: Has the submitting author (@freifrauvonbleifrei) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?"

I noticed on the [contributors page](https://github.com/SGpp/DisCoTec/graphs/contributors) that there seem to be 10 authors of the code (of which at least 4 seem to be major authors):
![image](https://github.com/user-attachments/assets/ae9b21f6-e427-4900-96d2-9686acd082aa)

but there are only 3 authors of the paper. I can identify the first 2 authors. Is @hurlerml Dirk Pflüger? Is it normal that Michael Obersteiner (@obersteiner ) is not listed as an author?
