### 2024-08-19 SGpp/DisCoTec#129 @EmilyBourne
[JOSS] Functionality documentation
Regarding https://github.com/openjournals/joss-reviews/issues/7018 - " Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?"

I cannot find documentation beyond the READMEs:
- https://github.com/SGpp/DisCoTec/blob/main/README.md
- https://github.com/SGpp/DisCoTec/blob/main/third_level_manager/README.md
- https://github.com/SGpp/DisCoTec/blob/main/tools/subspace_writer/README.md
- https://github.com/SGpp/DisCoTec/blob/main/examples/selalib_distributed/README.md

The webpage https://sparsegrids.org/ is mentioned in the "About" section but does not seem to mention DisCoTec:
![image](https://github.com/user-attachments/assets/cd54f0ff-3f77-401e-9508-64a677c845d1)

Is there API documentation describing the available classes/methods somewhere? Perhaps something generated with doxygen?
