### 2024-08-19 SGpp/DisCoTec#130 @EmilyBourne
[JOSS] State of the field
Regarding https://github.com/openjournals/joss-reviews/issues/7018 "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

The paper does not seem to mention other packages. Are the authors aware of any other packages for the sparse grid combination technique? Or other packages which could solve similar higher-dimensional problems to which this package could be compared?
