### 2024-08-19 SGpp/DisCoTec#131 @EmilyBourne
[JOSS] Paper query
Regarding https://github.com/openjournals/joss-reviews/issues/7018 I have a quick question. I don't know if there is a mistake in the paper or if I have simply misunderstood as I am unfamiliar with the sparse grid combination technique.

In the paper you state that the figure shows a scheme with 5 component grids but this does not seem to be reflected in the diagram. Depending on what is meant by a component grid I would expect there to be either 7, 4, or 3.
![image](https://github.com/user-attachments/assets/dd0f6926-52a6-4966-8684-d2e5678fdbb9)

Perhaps you meant to use the adjacent figure from the thesis:
![image](https://github.com/user-attachments/assets/9f0a1e8b-c132-4ea5-87b5-4385c8e92685)

?
