### 2024-08-21 SGpp/DisCoTec#128 @freifrauvonbleifrei
as far as I know, @hurlerml (a student contributor) is not at Uni Stuttgart any more, so I don't know if he'll react. But looking at the contributors page, I would agree that both @hurlerml and @obersteiner  should both at least have the chance to be authors.

@hurlerml @obersteiner would you like to? -> current paper draft is here: https://github.com/SGpp/DisCoTec/blob/feature_joss_submission/paper.md  (authors can be added in the metadata)
