### 2024-08-22 SGpp/DisCoTec#130 @freifrauvonbleifrei
I can quote from my thesis:

> The Sparse Grids Matlab Kit [126, 127] by Piazzola and Tamellini is targeted towards teaching and uncertainty quantification with the CT. It provides a user friendly MATLAB interface for the CT, and also dimensional adaptivity [53] for nested and non-nested sequences of component grid collocation points. The code’s parallelism is limited to shared memory only, as it does not target HPC applications. The same holds true for the sparseSpACE code [120], which provides an implementation for automatic spatial adaptivity in the CT for integration [121], interpolation, uncertainty quantification, machine learning methods, and PDE solvers.

I am not aware of something that targets distributed-memory combination technique.

@EmilyBourne @jakelangham what is your opinion, should we rather say "as far as we know there are no other codes for large-scale combination technique" or mention the above codes along with the limited comparability?
