### 2024-08-23 SGpp/DisCoTec#130 @EmilyBourne
In my opinion it would be good to mention the codes that are mentioned in your thesis. You can also say that there are no distributed-memory codes. This should clearly illustrate the gap that DisCoTec aims to fill
