### 2024-08-23 SGpp/DisCoTec#128 @hurlerml
Thank you for reaching out and for considering me as an author. I'm glad to hear that the code is being published to the Journal of Open Source Software. I would be happy to be mentioned. Please feel free to include me as an author.

Best regards,
Marcel

> Am 22.08.2024 um 01:45 schrieb Freifrau von Bleifrei ***@***.***>:
>
>
> as far as I know, @hurlerml <https://github.com/hurlerml> (a student contributor) is not at Uni Stuttgart any more, so I don't know if he'll react. But looking at the contributors page, I would agree that both @hurlerml <https://github.com/hurlerml> and @obersteiner <https://github.com/obersteiner> should both at least have the chance to be authors.
>
> @hurlerml <https://github.com/hurlerml> @obersteiner <https://github.com/obersteiner> would you like to? -> current paper draft is here: https://github.com/SGpp/DisCoTec/blob/feature_joss_submission/paper.md (authors can be added in the metadata)
>
