### 2024-08-24 #7018 @jakelangham
@danielskatz, @freifrauvonbleifrei I took a proper look today. (Apologies if I'm late.) From what I've seen at so far, it looks like a well written and potentially useful library that seems entirely suitable for a JOSS paper.

However, at the outset I'd like to echo @EmilyBourne's observation in SGpp/Discotec#129 about the lack of documentation. The README is nice and very welcome but I was expecting to see some sort of detailed guide for a newcomer about how to actually use the software. Without this, I cannot adequately review the library because I only have a partial understanding of what it does and what all its features are. Therefore, I think that major revisions are required in this direction before proceeding further.

The JOSS reviewer guidelines are somewhat discretionary on what counts as sufficient documentation, but essentially what I'm looking for is to convince myself that a target user could get to grips with DisCoTec and use it to solve problems completely independently of input from the developers and I don't think this is currently the case.
