### 2024-08-24 SGpp/DisCoTec#132 @jakelangham
[JOSS] Spack build failing
This issue relates to the ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

I was able to successfully build discotec via spack on two machines (a desktop and HPC).

However, on another cluster, the build fails to build one of the dependencies (libevent-2.1.12) because spack can't find openssl/ssl.h. This certainly isn't discotec's fault (it turns out that header is in a slightly odd place on that system), but ultimately it does affect the installation process. Therefore, perhaps for users like me that are unfamiliar with spack, you could add a brief bit of guidance to the README about how to get spack look for things in the right places. After all, it's pretty common when building on HPC to have to do this kind of troubleshooting. (This may also be helpful for the CMake installation option too, though people are probably more familiar with CMake in general.)
