### 2024-08-27 #7018 @freifrauvonbleifrei
very fair point about https://github.com/SGpp/DisCoTec/issues/129 . It may take a few weeks for us to get the documentation to a shareable state, I'll ping you when it's ready @jakelangham !
