### 2024-08-27 SGpp/DisCoTec#129 @pfluegdk
@EmilyBourne Thanks for pointing this out. DisCoTec was indeed missing in the list of software. This does not solve the API documentation issue, but I added DisCoTec on sparsegrids.org.
