### 2024-08-29 cdaed20
paper.md: add @hurlerml as author
