### 2024-08-29 SGpp/DisCoTec#128 @freifrauvonbleifrei
@hurlerml , I just saw your comment now, happy to hear from you! I just added you to the authors, let me know in case you want an ORCID added :)
