### 2024-08-30 480f3c6
Update paper.md
added @obersteiner to list of authors
