### 2024-09-02 2545934
docs: add basic sphinx configuration (generates with  doxygen Doxyfile && make html )
