+++ docs/index.rst
@@ -1,0 +1,24 @@
+.. DisCoTec documentation master file, created by
+   sphinx-quickstart on Thu Aug 29 13:22:11 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+DisCoTec's documentation!
+====================================
+
+This is about the C++ code for HPC simulations with the sparse grid combination technique.
+Not to be confused with the `International Federated Conference on Distributed Computing Techniques <http://www.discotec.org/>`!
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Contents:
+   
+   api/library_root
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
