### 2024-09-02 f007b3f
docs: add readthedocs config
