### 2024-09-03 b1efa01
docs: add reqirements.txt
