+++ docs/requirements.txt
@@ -1,0 +1,5 @@
+sphinx==7.2.6
+sphinx-rtd-theme==2.0.0
+breathe==4.35.0
+exhale==0.3.7
+myst-parser==2.0.0
