### 2024-09-12 76b2c8e
Task: default implementation for getFullGrid
