+++ src/task/Task.hpp
@@ -87,7 +87,7 @@ class Task {
 
   // returns the -th fullgrid gathered from all processors
   virtual void getFullGrid(FullGrid<CombiDataType>& fg, RankType lroot, CommunicatorType lcomm,
-                           int n = 0) = 0;
+                           int n = 0);
 
   // This method returns the local part of the n-th distributedFullGrid
   virtual DistributedFullGrid<CombiDataType>& getDistributedFullGrid(size_t n = 0) = 0;
