### 2024-09-12 33fea92
docs: add first tutorial draft
