### 2024-09-12 SGpp/DisCoTec#128 @freifrauvonbleifrei
Yay, with commit 65a03913b658370a9787987991dd854b8683960e we have added all major contributors!
