### 2024-09-12 6963be8
doc: add math display to conf.py
