+++ docs/advanced_topics.md
@@ -1,0 +1,101 @@
+# Advanced Topics and Further Resources
+
+## General Reading
+
+### Sparse Grids:
+- H.-J. Bungartz, M. Griebel. ‘Sparse grids’. In: Acta Numerica 13 (May 2004). Publisher: Cambridge University Press, pp. 147–269. url: https://www.cambridge.org/core/journals/acta-numerica/article/sparse-grids/47EA2993DB84C9D231BB96ECB26F615C
+- J. Garcke. ‘Sparse Grids in a Nutshell’. In: Sparse Grids and Applications. Ed. by J. Garcke, M. Griebel. Lecture Notes in Computational Science and Engineering. Berlin, Heidelberg: Springer, 2013, pp. 57–80.
+- D. Pflüger. Spatially Adaptive Sparse Grids for High-Dimensional Problems. Verlag Dr. Hut, 2010.
+- W. Guo, Y. Cheng. ‘A Sparse Grid Discontinuous Galerkin Method for High-Dimensional Transport Equations and Its Application to Kinetic Simulations’. In: SIAM Journal on Scientific Computing 38.6 (Jan. 1, 2016), A3381–A3409. url: https://epubs.siam.org/doi/10.1137/16M1060017
+- S. Schnake et al. ‘Sparse-grid discontinuous Galerkin methods for the Vlasov–Poisson–Lenard–Bernstein model‘. In: Journal of Computational Physics, vol. 510, p. 113053, Aug. 2024, doi: 10.1016/j.jcp.2024.113053.
+- A. Zeiser. ‘Sparse Grid Time-Discontinuous Galerkin Method with Streamline Diffusion for Transport Equations’. In: Partial Differential Equations and Applications 4.4 (Aug. 1, 2023), p. 38. url: https://doi.org/10.1007/s42985-023-00250-2 .
+
+
+### Combination Technique:
+- M. Griebel, M. Schneider, C. Zenger. ‘A combination technique for the solution of sparse grid problems’. In: Proceedings of the IMACS International Symposium on Iterative Methods in Linear Algebra: Brussels, Belgium, 2 - 4 April, 1991. Ed. by P. de Groen, R. Beauwens. North Holland, 1992.
+- M. Heene. ‘A Massively Parallel Combination Technique for the Solution of High-Dimensional PDEs’. PhD thesis. Institut für Parallele und Verteilte Systeme der Universität Stuttgart, 2018.
+
+
+# Variants of the Combination Technique in DisCoTec
+
+## Truncated Combination Technique
+DisCoTec by default uses the *truncated combination technique*, which sets a minimum level for all component grids used.
+
+- J. Benk, D. Pflüger. ‘Hybrid parallel solutions of the Black-Scholes PDE with the truncated combination technique’. In: 2012 International Conference on High Performance Computing Simulation (HPCS). 2012 International Conference on High Performance Computing Simulation (HPCS). July 2012.
+
+## Adaptivity
+
+Of the two ways of bringing adaptivity into the combination technique, DisCoTec currently supports only one: 
+Static dimensional adaptivity can be achieved by providing the combination scheme as a .json input file.
+These files can be generated with https://github.com/SGpp/DisCoTec-combischeme-utilities . 
+To get the more advanced combination schemes as discussed in C. Kowitz' dissertation, you may have to adapt the utilities code.
+If your application needs dynamic (= during runtime) or spatial adaptivity (= block-structured grids), please get in touch with the developers.
+
+- T. Gerstner, M. Griebel. ‘Dimension–adaptive tensor–product quadrature’. In: Computing 71.1 (2003), pp. 65–87.
+- C. Kowitz. ‘Applying the Sparse Grid Combination Technique in Linear Gyrokinetics’. Dissertation. München: Technische Universität München, 2016.
[…]
