+++ docs/advanced_topics.md
@@ -2,7 +2,7 @@
 
 ## General Reading
 
-### Sparse Grids:
+### Sparse Grids
 - H.-J. Bungartz, M. Griebel. ‘Sparse grids’. In: Acta Numerica 13 (May 2004). Publisher: Cambridge University Press, pp. 147–269. url: https://www.cambridge.org/core/journals/acta-numerica/article/sparse-grids/47EA2993DB84C9D231BB96ECB26F615C
 - J. Garcke. ‘Sparse Grids in a Nutshell’. In: Sparse Grids and Applications. Ed. by J. Garcke, M. Griebel. Lecture Notes in Computational Science and Engineering. Berlin, Heidelberg: Springer, 2013, pp. 57–80.
 - D. Pflüger. Spatially Adaptive Sparse Grids for High-Dimensional Problems. Verlag Dr. Hut, 2010.
@@ -11,19 +11,19 @@
 - A. Zeiser. ‘Sparse Grid Time-Discontinuous Galerkin Method with Streamline Diffusion for Transport Equations’. In: Partial Differential Equations and Applications 4.4 (Aug. 1, 2023), p. 38. url: https://doi.org/10.1007/s42985-023-00250-2 .
 
 
-### Combination Technique:
+### Combination Technique
 - M. Griebel, M. Schneider, C. Zenger. ‘A combination technique for the solution of sparse grid problems’. In: Proceedings of the IMACS International Symposium on Iterative Methods in Linear Algebra: Brussels, Belgium, 2 - 4 April, 1991. Ed. by P. de Groen, R. Beauwens. North Holland, 1992.
 - M. Heene. ‘A Massively Parallel Combination Technique for the Solution of High-Dimensional PDEs’. PhD thesis. Institut für Parallele und Verteilte Systeme der Universität Stuttgart, 2018.
 
 
-# Variants of the Combination Technique in DisCoTec
+## Variants of the Combination Technique in DisCoTec
 
-## Truncated Combination Technique
+### Truncated Combination Technique
 DisCoTec by default uses the *truncated combination technique*, which sets a minimum level for all component grids used.
 
 - J. Benk, D. Pflüger. ‘Hybrid parallel solutions of the Black-Scholes PDE with the truncated combination technique’. In: 2012 International Conference on High Performance Computing Simulation (HPCS). 2012 International Conference on High Performance Computing Simulation (HPCS). July 2012.
 
-## Adaptivity
+### Adaptivity
 
 Of the two ways of bringing adaptivity into the combination technique, DisCoTec currently supports only one: 
 Static dimensional adaptivity can be achieved by providing the combination scheme as a .json input file.
@@ -36,7 +36,7 @@ If your application needs dynamic (= during runtime) or spatial adaptivity (= bl
 - M. Obersteiner. ‘A spatially adaptive and massively parallel implementation of the fault-tolerant combination technique’. Dissertation. Technische Universität München, 2021. url: https://mediatum.ub.tum.de/doc/1613369/1613369.pdf .
 
 
-## Load Balancing
+### Load Balancing
 
 The static load balancing strategy used in the [tutorial](./simple_tutorial.md) is solely based on equally distributing the number of DOF, to allow for maximum usage of the main memory.
 This ignores that some grids may take a lot longer on certain operations due to their high anisotropy and the resulting cache effects.
@@ -47,7 +47,7 @@ To achieve anisotropy-based and dynamic load balancing, DisCoTec uses `LoadModel
[…]
