### 2024-09-17 71efacf
docs: add accidentally removed theme back in (rtd theme in conf.py)
