+++ docs/conf.py
@@ -27,6 +27,9 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
+html_theme = "sphinx_rtd_theme"
+html_static_path = ["_static"]
+
 
 # -- MyST configuration -----------------------------------------------------
 
