### 2024-09-17 5d1b38e
docu: move install etc. from README into docs
