+++ docs/index.rst
@@ -3,20 +3,28 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-DisCoTec's documentation!
-====================================
+DisCoTec
+========
 
-This is about the C++ code for HPC simulations with the sparse grid combination technique.
-Not to be confused with the `International Federated Conference on Distributed Computing Techniques <http://www.discotec.org/>`!
+This documentation is about the C++ code for HPC simulations with the sparse grid combination technique.
+Not to be confused with the `International Federated Conference on Distributed Computing Techniques <http://www.discotec.org/>`_!
 
 .. toctree::
    :maxdepth: 1
    :caption: Contents:
    
+   getting_started
    simple_tutorial
    advanced_topics
    api/library_root
 
+----------------
+READMEs
+----------------
+* `DisCoTec README <https://github.com/SGpp/DisCoTec/blob/main/README.md>`_ 
+* `Third level manager <https://github.com/SGpp/DisCoTec/blob/main/third_level_manager/README.md>`_
+* `Subspace writer tool <https://github.com/SGpp/DisCoTec/blob/main/tools/subspace_writer/README.md>`_
+* `SeLaLib Simulations with DisCoTec <https://github.com/SGpp/DisCoTec/blob/main/examples/selalib_distributed/README.md>`_
 
 Indices and tables
 ==================
