+++ docs/getting_started.md
@@ -1,0 +1,197 @@
+# Getting Started with DisCoTec
+
+## Installing DisCoTec
+
+There are two established ways of installing: Spack and CMake.
+After installing all dependencies, Spack actually delegates to
+CMake to build DisCoTec.
+
+### ... with Spack
+
+DisCoTec can be installed via spack, which handles all dependencies.
+
+If you want to use DisCoTec as a library only, follow the usual
+[Spack installation](https://spack.readthedocs.io/en/latest/features.html)
+instructions until the command `spack install discotec@main` 
+(or an extended version of it) succeeds.
+
+If you want to develop DisCoTec code or examples, the `spack dev-build` workflow
+is recommended as follows.
+
+Clone both `spack` and `DisCoTec` to find or build the dependencies and then
+compile DisCoTec:
+
+```bash
+git clone git@github.com:spack/spack.git  # use https if your ssh is not set up on github
+./spack/bin/spack external find  # find already-installed packages
+./spack/bin/spack compiler find  # find compilers present on system
+./spack/bin/spack info discotec@main  # shows DisCoTec's variants
+./spack/bin/spack spec discotec@main  # shows DisCoTec's dependency tree and which parts are already found
+
+git clone git@github.com:SGpp/DisCoTec.git
+cd DisCoTec
+../spack/bin/spack dev-build -b install discotec@main
+```
+
+This will first build all dependencies, and then build DisCoTec inside the
+cloned folder.
+The executables are placed in the respective `example` and `test` folders.
+
+If you encounter problems with the Spack installation, check out
+the [Spack command documentation](https://spack.readthedocs.io/en/latest/getting_started.html),
+and see if other users had the same 
+[issues](https://github.com/spack/spack/issues?q=is%3Aissue) before.
+
+### ... with CMake
+
+#### Dependencies
+
+cmake >= (3.24.2),
+libmpich-dev (>= 3.2-6), or other MPI library
+libboost-all-dev (>= 1.60)
+
+Additional (optional) dependencies:
+
+- OpenMP
+- HDF5
+- HighFive
+
+You can also install the dependencies with Spack, and `spack load`
+them, before executing the next steps.
+
+#### Complete build
+
+```bash
+git clone https://github.com/SGpp/DisCoTec.git DisCoTec
+cmake -S DisCoTec -B DisCoTec/build
+cmake --build DisCoTec/build -j
+```
+
+#### Advanced build
+
+All examples and the tools can be built on their own.
+To build a specific example, run the specific cmake command in the example folder.
+For example, run the following commands
+
+```bash
+git clone https://github.com/SGpp/DisCoTec.git DisCoTec
+cd DisCoTec/examples/combi_example
+cmake -S . -B build
+cmake --build build -j
+```
+
+to build the `combi_example`.
+
+For building only the DisCoTec library, run cmake with the `src` folder as
+source folder.
+
+#### Optional CMake Options
+
+- `DISCOTEC_TEST=**ON**|OFF` - Build tests if you build the complete project.
+- `DISCOTEC_BUILD_MISSING_DEPS=**ON**|OFF`- First order dependencies that are
+  not found are built automatically (glpk is always built).
+- `DISCOTEC_TIMING=**ON**|OFF` - Enables internal timing
+- `DISCOTEC_USE_HDF5=**ON**|OFF`
+- `DISCOTEC_USE_HIGHFIVE=**ON**|OFF` - Enables HDF5 support via HighFive. If
+  `DISCOTEC_USE_HIGHFIVE=ON`, `DISCOTEC_USE_HDF5` has also to be `ON`.
[…]
