+++ README.md
@@ -111,11 +111,8 @@ the currently open issues.
 
 ## Installing
 
-### Installation instructions: spack
-
 DisCoTec can be installed via spack, which handles all dependencies.
-If you want to develop DisCoTec code or examples, the `spack dev-build` workflow
-is recommended as follows.
+We recommend the `spack dev-build` workflow:
 
 Clone both `spack` and `DisCoTec` to find or build the dependencies and then
 compile DisCoTec:
@@ -136,172 +133,6 @@ This will first build all dependencies, and then build DisCoTec inside the
 cloned folder.
 The executables are placed in the respective `example` and `test` folders.
 
-### Installation instructions: CMake
+[Here are the Docs](https://discotec.readthedocs.io/en/latest/getting_started.html#installation-with-spack) for CMake options and further Spack customization hints.
 
-#### Dependencies
 
-cmake >= (3.24.2),
-libmpich-dev (>= 3.2-6), or other MPI library
-libboost-all-dev (>= 1.60)
-
-Additional (optional) dependencies:
-
-- OpenMP
-- HDF5
-- HighFive
-
-#### Complete build
-
-```bash
-git clone https://github.com/SGpp/DisCoTec.git DisCoTec
-cmake -S DisCoTec -B DisCoTec/build
-cmake --build DisCoTec/build -j
-```
-
-#### Advanced build
-
-All examples and the tools can be built on their own.
-To build a specific example, run the specific cmake command in the example folder.
-For example, run the following commands
-
-```bash
-git clone https://github.com/SGpp/DisCoTec.git DisCoTec
-cd DisCoTec/examples/combi_example
-cmake -S . -B build
-cmake --build build -j
-```
-
-to build the `combi_example`.
-
-For building only the DisCoTec library, run cmake with the `src` folder as
-source folder.
-
-#### Optional CMake Options
-
-- `DISCOTEC_TEST=**ON**|OFF` - Build tests if you build the complete project.
-- `DISCOTEC_BUILD_MISSING_DEPS=**ON**|OFF`- First order dependencies that are
-  not found are built automatically (glpk is always built).
-- `DISCOTEC_TIMING=**ON**|OFF` - Enables internal timing
-- `DISCOTEC_USE_HDF5=**ON**|OFF`
-- `DISCOTEC_USE_HIGHFIVE=**ON**|OFF` - Enables HDF5 support via HighFive. If
-  `DISCOTEC_USE_HIGHFIVE=ON`, `DISCOTEC_USE_HDF5` has also to be `ON`.
-- `DISCOTEC_UNIFORMDECOMPOSITION=**ON **|OFF` - Enables the uniform
-  decomposition of the grid.
-- `DISCOTEC_GENE=ON|**OFF**` - Currently GEne is not supported with CMake!
-- `DISCOTEC_OPENMP=ON|**OFF**` - Enables OpenMP support.
-- `DISCOTEC_ENABLEFT=ON|**OFF**` - Enables the use of the FT library.
-- `DISCOTEC_USE_LTO=**ON**|OFF` - Enables link time optimization if the compiler
-  supports it.
-- `DISCOTEC_OMITREADYSIGNAL=ON|**OFF**` - Omit the ready signal in the MPI
-  communication. This can be used to reduce the communication overhead.
-- `DISCOTEC_USENONBLOCKINGMPICOLLECTIVE=ON|**OFF**` - Flag currently unused
-- `DISCOTEC_WITH_SELALIB=ON|**OFF**` - Looks for SeLaLib dependencies and
-  compiles [the matching example](/examples/selalib_distributed/)
-
-To run the compiled tests, go to folder `tests` and run
-
-```bash
-mpiexec -np 9 ./test_distributedcombigrid_boost
-```
-
[…]
