### 2024-09-18 bf36617
docs: move parallelism to docs, extend README
