### 2024-09-18 8994d08
docs: adapt tutorial to alex' review
