### 2024-09-18 7c2ac59
docs: extend some advanced topics
