### 2024-09-18 #7018 @freifrauvonbleifrei
@danielskatz what a timely moment to ask :) The issue with the biggest impact was on feature documentation --
@obersteiner , @vancraar and me are currently putting the finishing touches on the new documentation at https://discotec.readthedocs.io/ !

At the same time, the main README is getting significantly shorter, see here: https://github.com/SGpp/DisCoTec/tree/feature_documentation .

@jakelangham maybe [this tutorial](https://discotec.readthedocs.io/en/latest/simple_tutorial.html) helps achieve
> that a target user could get to grips with DisCoTec and use it to solve problems completely independently of input from the developers

?
Let's continue this discussion in https://github.com/SGpp/DisCoTec/issues/129 :)
