### 2024-09-18 SGpp/DisCoTec#132 @freifrauvonbleifrei
the Spack installation is getting a bit more space in the new documentation: https://discotec.readthedocs.io/en/latest/getting_started.html , including links to the typical Spack resources. @jakelangham do you think you would know where to look based on this info? Unfortunately, resolving dependencies on HPC can be *very* complex, which is why we (and many HPC operators) try to outsource as much as possible to Spack and its documentation.
