### 2024-09-18 3bd92ac
docs: add boundary conditions
