### 2024-09-18 3cfe3d5
docs: put CT info into a doc page as well
