### 2024-09-18 a3ea59c
Create CODE_OF_CONDUCT.md
from contributor covenant
