### 2024-09-18 cd08fe2
readme: move contributing to own file
