### 2024-09-18 e22b399
Contributing: expand a section
