### 2024-09-18 SGpp/DisCoTec#127 @freifrauvonbleifrei
As part of the README / Docs restructuring, I bundled this CONTRIBUTING.md file:
<https://github.com/SGpp/DisCoTec/blob/feature_documentation/CONTRIBUTING.md>
Let me know what you think!
