### 2024-09-18 #7018 @jakelangham
@freifrauvonbleifrei Fantastic - I'm away this week, so will take a proper look when it's fully done. At first glance it looks like exactly what we're after.
