### 2024-09-25 13cb2ae
docs: incorporate some of alex' feedback
