+++ docs/getting_started.md
@@ -22,13 +22,13 @@ Clone both `spack` and `DisCoTec` to find or build the dependencies and then
 compile DisCoTec:
 
 ```bash
-git clone git@github.com:spack/spack.git  # use https if your ssh is not set up on github
+git clone https://github.com/spack/spack.git  # use https if your ssh is not set up on github
 ./spack/bin/spack external find  # find already-installed packages
 ./spack/bin/spack compiler find  # find compilers present on system
 ./spack/bin/spack info discotec@main  # shows DisCoTec's variants
 ./spack/bin/spack spec discotec@main  # shows DisCoTec's dependency tree and which parts are already found
 
-git clone git@github.com:SGpp/DisCoTec.git
+git clone https://github.com/SGpp/DisCoTec.git
 cd DisCoTec
 ../spack/bin/spack dev-build -b install discotec@main
 ```
@@ -40,7 +40,8 @@ The executables are placed in the respective `example` and `test` folders.
 If you encounter problems with the Spack installation, check out
 the [Spack command documentation](https://spack.readthedocs.io/en/latest/getting_started.html),
 and see if other users had the same 
-[issues](https://github.com/spack/spack/issues?q=is%3Aissue) before.
+[spack issues](https://github.com/spack/spack/issues?q=is%3Aissue) or 
+[DisCoTec issues](https://github.com/SGpp/DisCoTec/issues) before.
 
 ### ... with CMake
 
@@ -48,13 +49,16 @@ and see if other users had the same
 
 cmake >= (3.24.2),
 libmpich-dev (>= 3.2-6), or other MPI library
-libboost-all-dev (>= 1.60)
+libboost-all-dev (>= 1.60) (or the subset test, serialization, filesystem, system, program_options, date_time Boost libraries)
 
 Additional (optional) dependencies:
 
 - OpenMP
 - HDF5
 - HighFive
+- lz4
+- glpk (used version bundled in the DisCoTec repo)
+- vtk 
 
 You can also install the dependencies with Spack, and `spack load`
 them, before executing the next steps.
@@ -115,13 +119,12 @@ To run the compiled tests, go to folder `tests` and run
 mpiexec -np 9 ./test_distributedcombigrid_boost
 ```
 
-where you can use all the parameters of the boost test suite.
+where you can use all the [parameters of the boost test suite](https://beta.boost.org/doc/libs/1_60_0/libs/test/doc/html/boost_test/utf_reference/rt_param_reference.html).
 If timings matter, consider [pinning](#pinning-with-various-mpi-implementations).
-Or you can run the tests with `ctest` in the build folder.
 
 ## Run an Example
 
-DisCoTec executables are typicall configured through `ctparam` files, which are parsed on
+DisCoTec executables are typically configured through `ctparam` files, which are parsed on
 startup.
 The `ctparam` file will contain the combination technique parameters (dimension,
 minimum and maximum level) as well as parallelization parameters (number and
