### 2024-09-25 75a0f0a
added fault tolerance description
