### 2024-09-26 12f0431
paper.md: format, small edits
(replace some "offers")
