### 2024-09-26 667fa91
paper.bib: add matlab kit reference
