+++ paper.bib
@@ -84,3 +84,17 @@
   isbn = {978-1-4503-3723-6},
   langid = {english}
 }
+
+@online{piazzolaSparseGridsMatlab2022,
+  title = {The {{Sparse Grids Matlab}} Kit -- a {{Matlab}} Implementation of Sparse Grids for High-Dimensional Function Approximation and Uncertainty Quantification},
+  author = {Piazzola, Chiara and Tamellini, Lorenzo},
+  date = {2022-03-17},
+  eprint = {2203.09314},
+  eprinttype = {arXiv},
+  eprintclass = {cs, math},
+  doi = {10.48550/arXiv.2203.09314},
+  url = {http://arxiv.org/abs/2203.09314},
+  urldate = {2023-03-15},
+  abstract = {The Sparse Grids Matlab kit is a collection of Matlab functions for high-dimensional interpolation and quadrature, based on the combination technique form of sparse grids. It is lightweight, high-level and easy to use, good for quick prototyping and teaching. It is somehow geared towards Uncertainty Quantification (UQ), but it is flexible enough for other purposes. The goal of this paper is to give an overview of the implementation structure of the Sparse Grids Matlab kit and to showcase its potentialities, guided by some illustrative tests and a final comprehensive example on forward and inverse UQ analysis.},
+  keywords = {Computer Science - Mathematical Software,Mathematics - Numerical Analysis},
+}
