### 2024-09-26 7596256
paper.md: reference SG matlab kit
