### 2024-09-26 SGpp/DisCoTec#130 @freifrauvonbleifrei
hi @EmilyBourne , there is a new State of the Field section in the [current paper version](https://github.com/SGpp/DisCoTec/actions/runs/11043506452/artifacts/1979893307)!
