### 2024-09-26 SGpp/DisCoTec#129 @freifrauvonbleifrei
@EmilyBourne @jakelangham  I believe the documentation is fine for review! ->  https://discotec.readthedocs.io/

with a few edits pending, which can be followed in https://github.com/SGpp/DisCoTec/pull/133

Please do let us know if things are unclear, your "outside" perspective can really help us make this a great documentation :)
