### 2024-09-26 a79a42b
Update advanced_topics.md with Data transfer for Widely-Distributed Simulations
