### 2024-09-26 3402c14
docs: formatting in markdown (mostly codacy suggestions)
