+++ docs/simple_tutorial.md
@@ -4,7 +4,9 @@ So you want to use your timestepped simulation with the DisCoTec framework.
 This Tutorial gives you an outline of the steps required.
 
 ## Your Code Interface: Init, Timestep, Get/Set Full Grid, Finish
-The first step is to prepare your code in a way that it can be called by DisCoTec.
+
+The first step is to prepare your code in a way that it can be called by
+DisCoTec.
 Typically, your simulation code will look similar to this (in pseudocode):
 
 ```python
@@ -14,7 +16,14 @@ int num_points_z = ...
 int num_points_vx = ...
 ...
 initial_function = ...
-grid.initialize(initial_function, num_points_x, num_points_y, num_points_z, num_points_vx, ... )
+grid.initialize(
+                initial_function,
+                num_points_x,
+                num_points_y,
+                num_points_z,
+                num_points_vx,
+                ... 
+               )
 helper_data_structures.initialize(...)
 float end_time = ...
 float time_step = ...
@@ -29,14 +38,20 @@ properties = compute_properties(grid)
 write_output(properties, grid)
 ```
 
-For use with DisCoTec, we assume nestable power-of-two discretizations, i.e. where your grid spacing can be $2^{-l}$ for $l \in \mathbb{N}$.
-For simplicity, in this tutorial we also assume you use periodic boundary conditions and all `num_points_` are chosen as powers of two.
+For use with DisCoTec, we assume nestable power-of-two discretizations, i.e.,
+where your grid spacing can be $2^{-l}$ for $l \in \mathbb{N}$.
+For simplicity, in this tutorial we also assume you use periodic boundary
+conditions and all `num_points_` are chosen as powers of two.
 
-You need to transform your current solver code into stateful functions, or a stateful data structure.
-Let's say you introduce a class `YourSimulation`: computations before the time loop go into its constructor 
-or an `init()` function (if your code uses MPI, this is also where a sub-communicator should be passed for `YourSimulation` to use).
-Computation in the time loop goes into its `run()` function. 
-Anything after the time loop goes into the destructor or a function `finalize()`.
+You need to transform your current solver code into stateful functions, or a
+stateful data structure.
+Let's say you introduce a class `YourSimulation`: computations before the time
+loop go into its constructor
+or an `init()` function (if your code uses MPI, this is also where a
+sub-communicator should be passed for `YourSimulation` to use).
+Computation in the time loop goes into its `run()` function.
+Anything after the time loop goes into the destructor or a function
+`finalize()`.
 Then, the following should do the same:
 
 ```diff
@@ -47,9 +62,22 @@ int num_points_vx = ...
 ...
 
 - initial_function = ...
-- grid.initialize(initial_function, num_points_x, num_points_y, num_points_z, num_points_vx, ... )
+- grid.initialize(
+-                 initial_function,
+-                 num_points_x,
+-                 num_points_y,
+-                 num_points_z,
+-                 num_points_vx,
+-                 ... 
+-                )
 - helper_data_structures.initialize(...)
-+ my_simulation = YourSimulation(num_points_x, num_points_y, num_points_z, num_points_vx, ...)
++ my_simulation = YourSimulation(
++                                num_points_x, 
++                                num_points_y,
++                                num_points_z,
++                                num_points_vx,
++                                ...
++                               )
 float end_time = ...
 float time_step = ...
 float time_now = 0.0;
@@ -65,14 +93,18 @@ while(time_now < end_time) { # time loop
 + my_simulation.finalize()
 ```
 
-This setup assumes that we can pass the number of grid points per dimension in your high-d contiguous array as arguments.
-In addition, you will need an interface that allows DisCoTec to get and set this data.
[…]
