+++ docs/advanced_topics.md
@@ -3,140 +3,240 @@
 ## General Reading
 
 ### Sparse Grids
-- H.-J. Bungartz, M. Griebel. ‘Sparse grids’. In: Acta Numerica 13 (May 2004). Publisher: Cambridge University Press, pp. 147–269. url: https://www.cambridge.org/core/journals/acta-numerica/article/sparse-grids/47EA2993DB84C9D231BB96ECB26F615C
-- J. Garcke. ‘Sparse Grids in a Nutshell’. In: Sparse Grids and Applications. Ed. by J. Garcke, M. Griebel. Lecture Notes in Computational Science and Engineering. Berlin, Heidelberg: Springer, 2013, pp. 57–80.
-- D. Pflüger. Spatially Adaptive Sparse Grids for High-Dimensional Problems. Verlag Dr. Hut, 2010.
-- W. Guo, Y. Cheng. ‘A Sparse Grid Discontinuous Galerkin Method for High-Dimensional Transport Equations and Its Application to Kinetic Simulations’. In: SIAM Journal on Scientific Computing 38.6 (Jan. 1, 2016), A3381–A3409. url: https://epubs.siam.org/doi/10.1137/16M1060017
-- S. Schnake et al. ‘Sparse-grid discontinuous Galerkin methods for the Vlasov–Poisson–Lenard–Bernstein model‘. In: Journal of Computational Physics, vol. 510, p. 113053, Aug. 2024, doi: 10.1016/j.jcp.2024.113053.
-- A. Zeiser. ‘Sparse Grid Time-Discontinuous Galerkin Method with Streamline Diffusion for Transport Equations’. In: Partial Differential Equations and Applications 4.4 (Aug. 1, 2023), p. 38. url: https://doi.org/10.1007/s42985-023-00250-2 .
 
+- H.-J. Bungartz, M. Griebel. ‘Sparse grids’. In: Acta Numerica 13 (May 2004).
+  Publisher: Cambridge University Press, pp. 147–269. url: <https://www.cambridge.org/core/journals/acta-numerica/article/sparse-grids/47EA2993DB84C9D231BB96ECB26F615C>.
+- J. Garcke. ‘Sparse Grids in a Nutshell’. In: Sparse Grids and Applications.
+  Ed. by J. Garcke, M. Griebel. Lecture Notes in Computational Science and Engineering.
+  Berlin, Heidelberg: Springer, 2013, pp. 57–80.
+- D. Pflüger. Spatially Adaptive Sparse Grids for High-Dimensional Problems.
+  Verlag Dr. Hut, 2010.
+- W. Guo, Y. Cheng. ‘A Sparse Grid Discontinuous Galerkin Method for High-Dimensional
+  Transport Equations and Its Application to Kinetic Simulations’.
+  In: SIAM Journal on Scientific Computing 38.6 (Jan. 1, 2016), A3381–A3409.
+  url: <https://epubs.siam.org/doi/10.1137/16M1060017>.
+- S. Schnake et al. ‘Sparse-grid discontinuous Galerkin methods for the
+  Vlasov–Poisson–Lenard–Bernstein model‘. In: Journal of Computational Physics,
+  vol. 510, p. 113053, Aug. 2024, doi: 10.1016/j.jcp.2024.113053.
+- A. Zeiser. ‘Sparse Grid Time-Discontinuous Galerkin Method with Streamline
+  Diffusion for Transport Equations’.
+  In: Partial Differential Equations and Applications 4.4 (Aug. 1, 2023), p. 38.
+  url: <https://doi.org/10.1007/s42985-023-00250-2>.
 
 ### Combination Technique
[…]
