### 2024-09-26 06784c4
docs: some more markdown formatting (codacy suggestions)
