### 2024-09-26 a127029
docs: format sub-readmes
