### 2024-09-26 SGpp/DisCoTec#132 @jakelangham
I think the new documentation is sufficient to close this.

I got the code built on all 3 systems in the end with spack and cmake, so that feels like a decent test of the install process. FYI, I actually didn't find the spack documentation all that useful in this regard, since it seems more aimed at package developers and sysadmins than impatient users who just want to build 1 piece of software. But that is hardly your responsibility!
