### 2024-09-26 SGpp/DisCoTec#129 @freifrauvonbleifrei
> Further to @EmilyBourne's initial comment - there should be a link to the new docs somewhere in the README.md that's at least as prominent as the sparsegrids webpage link.

I am trying something -- can you spot it?
