### 2024-09-26 SGpp/DisCoTec#129 @freifrauvonbleifrei
> It is up to you how you do this but one suggestion would be to add some more text to the 'What is Discotec?' section that explains/advertises the benefits of using it and perhaps suggests the sorts of applications it might be particular suited for.

@jakelangham  I thought that the ["When (not) to use DisCoTec" sections](https://github.com/SGpp/DisCoTec/tree/feature_documentation?tab=readme-ov-file#when-to-use-discotec) could fulfill this function, but if you didn't find it, it's obviously not prominent enough yet (or not what you were looking for?). Maybe it would be more logical to put above the installation section anyways...
