### 2024-09-27 SGpp/DisCoTec#129 @jakelangham
@freifrauvonbleifrei Ah, I had not fully appreciated yesterday that I should be reviewing the feature_documentation branch... What you've written looks ideal and just what I was looking for. It could go a bit higher up in the README but that's up to you.

p.s. It might be a requirement for JOSS that these commits are merged into main before publication.
