### 2024-09-27 b04ef1d
README: fix dead link in spack badge
cf. https://github.com/spack/spack/issues/46604
