### 2024-09-27 b822ddf
README: reordering
cf. https://github.com/SGpp/DisCoTec/issues/129#issuecomment-2378121051
