### 2024-09-28 SGpp/DisCoTec#134 @jakelangham
[JOSS] Fig 1
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

I have been going through the README and new documentation and got a bit stuck on the first figure. I think the idea is _just about_ clear but it is not presented in a self-contained way and the writing is very terse. (I can see why that might be a good thing in the README but in the documentation you are free to go into detail.) There is a lot of notation in the figure that is not defined so even though I think I've guessed correctly what e.g. l_x and l_y are supposed to mean, I can't be exactly sure.

I've since looked at the paper and here there is a more detailed explanation of the figure, which helped me, though still there is left undefined.

Finally, I think a few brief comment on the properties of the example scheme might be helpful. I suppose the implication of the figure is that the sparse grids combine to give an effective discretisation indicated by the black-and-white diagram in the top right. Can you say anything about how this would compare to discretising over an 'equivalent' full grid (whose resolution was given by the finest scale present in the sparse grids)? Would the sparsely-combined method have the same order of accuracy? Would it be more efficient, or do you need to go to higher dimensional problems to see the benefits? Perhaps this all depends on details outside the scope of the example, but even so, anything you could say could be useful for a reader less familiar with the methods than yourself.
