### 2024-09-29 SGpp/DisCoTec#135 @jakelangham
[JOSS] SGppDistributedCombigridModule test failing
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

I have been trying to get the tests working on a cluster I use. I've done a fresh install with spack, following the documentation and submitted with the attached script.

```
#!/bin/bash
#SBATCH --job-name=discotest
#SBATCH --mem-per-cpu 8000M
#SBATCH --nodes=1
#SBATCH --ntasks-per-node=9
#SBATCH --cpus-per-task=1
#SBATCH --time=01-00:00:0

cd $SLURM_SUBMIT_DIR

[path-to-mpiexec] -np 9 ./test_distributedcombigrid_boost
```

where path-to-mpiexec refers to the binary that spack locally installed. (N.b. the install documentation assumes that mpiexec is already in $PATH which I think would not be true by default if following the spack route.)

While some tests seem to run ok, it eventually outptuts some rather formidable error messages, purportedly related to the ```SGppDistributedCombigridModule```. I attach the output file in the hope that you can make sense of it [slurm-10530966.txt](https://github.com/user-attachments/files/17178770/slurm-10530966.txt). Perhaps this could be a local problem with my install / setup but I'm not exactly sure how to fix it.
