### 2024-09-30 SGpp/DisCoTec#135 @freifrauvonbleifrei
Hi @jakelangham , thanks for testing the tests!

Yes, I agree that we should mention mpiexec has to be the one belonging to the MPI used in Spack.

Yes, this last set of tests! they are related to the widely-distributed combination technique, based on file exchange.
I believe there is some kind of race condition happening due to lacking synchronzation through the file system (only sometimes). We also see them occasionally in the Jenkins CI. The error you linked sounds like some outdated file is read (which is not large enough).

Right now, I am unsure how to go about this (debugging could be quick and just need a barrier, or it could take ages).

could you maybe delete the temporary files and run only
`${mpiexec} -n 9 ./test_distributedcombigrid_boost --run_test=thirdLevel --log_level=message`  to see which exact one it is on your machine?
