### 2024-09-30 SGpp/DisCoTec#135 @jakelangham
Absolutely - see the attached log

[slurm-10531195.txt](https://github.com/user-attachments/files/17186541/slurm-10531195.txt)
