### 2024-09-30 SGpp/DisCoTec#137 @jakelangham
[JOSS] paper suggestions
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

I have read through the draft manuscript. Overall I thought it was well written and fits the style of a JOSS paper. In addition to a few minor things included in this pr https://github.com/SGpp/DisCoTec/pull/136, I have a few suggestions listed below

- A reference or two for the 'sparse grid combination technique' would be useful early on.
- In the summary you say 'DisCoTec demonstrates its superiority' without making it clear what you are comparing it with. Moreover, does 'superiority' refer to accuracy, resource use, efficiency, etc?
- Acronyms (such as HPC, UQ etc) should be defined before use.
- The 'Statement of need' would benefit from references that cover the curse of dimensionality and the approaches that attempt to deal with it.
- When you say 'multi-scale methods ... provide an alternative approach to address the curse of dimensionality' - if you can do so compactly, it would be good to explain how they achieve this.
- When you say 'some other implementations of the sparse grid combination technique are available' - references should be provided if possible.
- The second paragraph of the 'Statement of need' largely repeats material from the 'Summary' and could (in my opinion) be deleted.
- I mentioned this already in https://github.com/SGpp/DisCoTec/issues/134, but care needs to be taken to define all mathematical notation. For example, in the combination formula, $f^{(s)}$, $f_{\vec{\ell}}$ and $c_{\vec{\ell}}$ are defined, but not $\vec{\ell}$ or $\mathcal{I}$.
- This doesn't necessarily need to go in the paper, but could you explain how using OpenMP decreases the memory footprint?
- In a couple of places, you provide links to github repositories that could (and perhaps should) be made into full references in the bibliography.
- When mentioning spack - you could indicate that the purpose of this is to simplify installation on HPC systems.
- On line 82: 'The C++ code SG++ allows to...' - allows what/who to?
