### 2024-10-01 SGpp/DisCoTec#135 @freifrauvonbleifrei
Hm, this output looks like the test suite just didn't finish in time, which could be fixed by a longer test timeout ;)

But this test suite is really volatile because of the files, I am considering to default-disable them in addition, but leave them in the CI so we know in case they REALLY break. What do you think @jakelangham ?
