### 2024-10-01 SGpp/DisCoTec#135 @jakelangham
I don't have a problem with that, except I'd like to verify that I can get them to work as expected first. Is there a header file or parameter somewhere I can tweak for this?
