### 2024-10-08 #7018 @EmilyBourne
@danielskatz I am in a similar place to @jakelangham .

I was waiting for the documentation to be addressed so I could try to use the software on some of my own problems to ensure that the functionality is ok and that the documentation is clear enough to make this possible.
However for the moment I agree with @jakelangham 's comment on https://github.com/SGpp/DisCoTec/issues/129

> I was just about able to follow the tutorial, which I think is very valuable. Nevertheless, to understand the code snippets I often wanted to see more context on what some of the functions called are doing. This led me to click on the 'Library API', which is yet to exist. I hate to say it, but this could be useful. API documentation is listed as a requirement in the JOSS review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) and seems particularly important for a code such as this which is designed to work like a library that users build codes on top of.

Without a Library API it is difficult to have an overview of the functionalities. We can check the ones used in the tutorial but I imagine that these are not the only functions available.

I am also waiting for https://github.com/SGpp/DisCoTec/pull/133 to be merged before checking the associated boxes in my review.
