### 2024-10-16 SGpp/DisCoTec#135 @freifrauvonbleifrei
Sure! in this file https://github.com/SGpp/DisCoTec/blob/main/tests/test_thirdLevel.cpp you can search for `boost::unit_test::timeout` . The first one you find, on line 861, is for the whole test suite and the others are for the individual tests. You can also remove them entirely, if you remember to stop the process at some point in case it hangs ;)

Before you run again, make sure to run `git clean -n` and delete all untracked files in the `tests` folder, to avoid interference from previously aborted runs' files.
