### 2024-10-16 6ae8e4a
docs: add mpi loading to getting_started cf. https://github.com/SGpp/DisCoTec/issues/135
