### 2024-10-16 ad16b81
docs: make doxygen output path findable (?)
