### 2024-10-21 #7018 @danielskatz
👋 @freifrauvonbleifrei - Can you update us on your progress (and/or schedule) for responding to the issues from @jakelangham and @EmilyBourne?
