### 2024-10-22 #7018 @freifrauvonbleifrei
@danielskatz thank you for raising the uncomfortable questions! ;)

I just figured that despite https://github.com/SGpp/DisCoTec/commit/ad16b810d2ec1cf8cc31eb00d76f8e47ebe83827 , the Doxygen paths are not quite right yet, but that should be an easier thing to fix.

The harder questions were arising for me once I started fine-grained documentation: I don't think I and the co-authors will be able to cover the full API in detail -- this is the curse of a multi-PhD-generational code (it's broad, and would need significant brainpower to re-understand every part). But I'd rather not throw out potentially useful functionality. On the other hand, snippets of Doxygen are sprinkled throughout all parts of the code, making the current Doxygen overview quite messy to read.

So my suggestion would be to focus on the classes that framework users would most likely interact with (`DistributedSparseGridUniform`, `ProcessGroupWorker`, `TaskWorker`, `SparseGridWorker`, `CombiParameters`, `CombiScheme`, `CombiCom`, `ProcessManager` and `ProcessGroupManager`, in my opinion).
Once this is done (by the end of next week?), I would merge https://github.com/SGpp/DisCoTec/pull/133 ; so that @EmilyBourne and @jakelangham could start on integrating it with their applications, and let me know in which parts more or less documentation depth is required.
(Concurrently, I would try to understand Doxygen sufficiently well to remove the "Library API" documentation parts that are obviously useless.)

Please do let me know your thoughts and inner resistances -- quite possibly, there is a smarter way to go about it.
