### 2024-10-22 3ceaf94
docs: make doxygen output path findable (??)
