### 2024-10-22 #7018 @jakelangham
@freifrauvonbleifrei I agree with this approach - it's best to focus on the core functionality. Even better if you can highlight the important parts of the API (or remove the unnecessary auto-generated boilerplate stuff). The review criteria for JOSS indicate that only the core functions/classes needed to be documented (though obviously more documentation is usually better).
