### 2024-10-23 faa2b5a
docs: fix broken relative links
