### 2024-10-25 2c5b728
Task: document and format
