+++ src/task/Task.hpp
@@ -6,29 +6,58 @@
 #include <boost/serialization/vector.hpp>
 #include <string>
 #include <vector>
+
 #include "fault_tolerance/FaultCriterion.hpp"
 #include "fault_tolerance/StaticFaults.hpp"
 #include "fullgrid/DistributedFullGrid.hpp"
 #include "fullgrid/FullGrid.hpp"
+#include "loadmodel/LoadModel.hpp"
 #include "mpi/MPISystem.hpp"
 #include "utils/LevelVector.hpp"
-#include "loadmodel/LoadModel.hpp"
-
 
 namespace combigrid {
 
 /**
- * Interface for tasks
+ * @brief Abstract class for a task that is worked on by a process group (consisting of
+ * ProcessGroupWorkers)
+ *
+ * In the combination technique, this will correspond to the current solution of the PDE for a
+ * component grid of a certain level, along with the operators to advance further in time.
+ *
+ * The task's domain is assumed to be the unit hypercube ((0,1)^d or [0,1)^d [0,1]^d , depending on
+ * the boundary conditions)
  */
-
 class Task {
  protected:
+  /**
+   * @brief Task default constructor, required for serialization
+   */
   Task();
 
+  /**
+   * @brief Task constructor
+   *
+   * @param dim dimensionality of the task's domain
+   * @param l levelvector of the task
+   * @param boundary boundary conditions of the task
+   * @param coeff combination coefficient of the task
+   * @param loadModel load model for the task, required for dynamic load balancing
+   * @param faultCrit fault criterion for the task, required for fault simulation
+   */
   Task(DimType dim, const LevelVector& l, const std::vector<BoundaryType>& boundary, real coeff,
        LoadModel* loadModel,
        FaultCriterion* faultCrit = (new StaticFaults({0, IndexVector(0), IndexVector(0)})));
 
+  /**
+   * @brief Task constructor
+   *
+   * @param l levelvector of the task
+   * @param boundary boundary conditions of the task
+   * @param coeff combination coefficient of the task
+   * @param loadModel load model for the task, required for dynamic load balancing
+   * @param faultCrit fault criterion for the task, required for fault simulation
+   *
+   */
   Task(const LevelVector& l, const std::vector<BoundaryType>& boundary, real coeff,
        LoadModel* loadModel,
        FaultCriterion* faultCrit = (new StaticFaults({0, IndexVector(0), IndexVector(0)})));
@@ -44,85 +73,200 @@ class Task {
 
  public:
   virtual ~Task();
-  // pgroup receive task
+
+  /**
+   * @brief receive a task from another rank
+   */
   static void receive(Task** t, RankType source, CommunicatorType comm);
 
-  // manager send task to pgroup root
+  /**
+   * @brief send a task to another rank
+   */
   static void send(const Task* const t, RankType dest, CommunicatorType comm);
 
-  // broadcast task
+  /**
+   * @brief broadcast a task to all ranks (typically within the same process group)
+   */
   static void broadcast(Task** t, RankType root, CommunicatorType comm);
 
+  /**
+   * @brief get the dimensionality of the task's domain
+   */
   inline DimType getDim() const;
 
+  /**
+   * @brief get the level vector of the task
+   */
   inline const LevelVector& getLevelVector() const;
 
+  /**
+   * @brief get the unique ID of the task
+   */
   inline size_t getID() const;
 
+  /**
+   * @brief get the combination coefficient of the task
+   */
   inline real getCoefficient() const;
 
   /**
    * @brief explicitly set a new ID for the task;
-   * may be useful if process groups read their task assignment from file instead of receiving it
+   *
+   * useful if process groups read their task assignment from file instead of receiving it
    * from the manager
    *
[…]
