### 2024-10-25 02100d0
dfg: reorder fcn definitions to match declarations
