+++ src/fullgrid/DistributedFullGrid.hpp
@@ -286,7 +286,8 @@ class DistributedFullGrid {
   /** get coordinates of rank r's upper bounds */
   inline std::vector<real> getUpperBoundsCoords(RankType r) const;
 
-  /* get the neighboring process' rank in a given dimension
+  /**
+   * @brief get the neighboring process' rank in a given dimension
    *
    * in the sense that the neighbor has the same
    * partion coordinates in all other dimensions than d; and
@@ -665,155 +666,6 @@ DistributedFullGrid<FG_ELEMENT>::~DistributedFullGrid() {
   }
 }
 
-template <typename FG_ELEMENT>
-double DistributedFullGrid<FG_ELEMENT>::getPointDistanceToCoordinate(
-    IndexType oneDimensionalLocalIndex, double coord, DimType d) const {
-  const auto& h = getGridSpacing();
-  auto coordDistance =
-      static_cast<double>(this->getLowerBounds()[d] + (hasBoundaryPoints_[d] > 0 ? 0 : 1) +
-                          oneDimensionalLocalIndex) *
-          h[d] -
-      coord;
-#ifndef NDEBUG
-  if (std::abs(coordDistance) > h[d]) {
-    std::cout << "assert bounds " << coordDistance << coord << h << static_cast<int>(d)
-              << oneDimensionalLocalIndex << std::endl;
-    assert(false &&
-           "should only be called for coordinates within the support of this point's basis "
-           "function");
-  }
-#endif  // ndef NDEBUG
-  return std::abs(coordDistance);
-}
-
-template <typename FG_ELEMENT>
-FG_ELEMENT DistributedFullGrid<FG_ELEMENT>::evalIndexAndAllUpperNeighbors(
-    const IndexVector& localIndex, const std::vector<real>& coords) const {
-  FG_ELEMENT result = 0.;
-  auto localLinearIndex = this->getLocalLinearIndex(localIndex);
-  for (size_t localIndexIterate = 0;
-       localIndexIterate < combigrid::powerOfTwoByBitshift(this->getDimension());
-       ++localIndexIterate) {
-#ifndef NDEBUG
-    auto neighborVectorIndex = localIndex;
-#endif
-    auto neighborIndex = localLinearIndex;
-    real phi_c = 1.;  // value of product of iterate's basis function on coords
-    for (DimType d = 0; d < dim_; ++d) {
-      const auto lastIndexInDim = this->getLocalSizes()[d] - 1;
-      bool isUpperInDim = std::bitset<sizeof(int) * CHAR_BIT>(localIndexIterate).test(d);
-      auto iterateIndexInThisDimension = localIndex[d] + isUpperInDim;
-      if (iterateIndexInThisDimension < 0) {
-        phi_c = 0.;
-        break;
-      } else if (isUpperInDim && this->hasBoundaryPoints_[d] == 1 &&
-                 //  this->getCartesianUtils().isOnLowerBoundaryInDimension(d) &&
-                 iterateIndexInThisDimension == this->getGlobalSizes()[d]) {
-        // if we have a wrap-around AND are on the lowest cartesian process in this dimension
-        // we need to set the index in this dim to 0 and shift the coordinate by 1.
-        neighborIndex -= localIndex[d] * this->getLocalOffsets()[d];
-        iterateIndexInThisDimension = 0;
-        auto coordDistance =
-            getPointDistanceToCoordinate(iterateIndexInThisDimension, coords[d] - 1.0, d);
-        auto oneOverHinD = powerOfTwo[levels_[d]];
-        phi_c *= 1. - coordDistance * oneOverHinD;
[…]
