### 2024-10-25 2a16861
docs: add documentation for sparse grid classes
