+++ src/sparsegrid/DistributedSparseGridUniform.hpp
@@ -1,15 +1,14 @@
 #ifndef SRC_SGPP_COMBIGRID_SPARSEGRID_DISTRIBUTEDSPARSEGRIDUNIFORM_HPP_
 #define SRC_SGPP_COMBIGRID_SPARSEGRID_DISTRIBUTEDSPARSEGRIDUNIFORM_HPP_
-#include "sparsegrid/AnyDistributedSparseGrid.hpp"
-#include "utils/LevelSetUtils.hpp"
-#include "utils/Types.hpp"
-
 #include <boost/iterator/counting_iterator.hpp>
 #include <cassert>
 #include <numeric>
 #include <set>
 #include <vector>
 
+#include "sparsegrid/AnyDistributedSparseGrid.hpp"
+#include "utils/LevelSetUtils.hpp"
+#include "utils/Types.hpp"
 
 namespace combigrid {
 // forward declarations
@@ -19,6 +18,12 @@ class DistributedFullGrid;
 template <typename FG_ELEMENT>
 class DistributedSparseGridUniform;
 
+/**
+ * @brief Container for distributed sparse grid data
+ *
+ * This class is used to store the data of a distributed sparse grid in a contiguous memory block.
+ * In addition, it provides storage for Kahan summation terms.
+ */
 template <typename FG_ELEMENT>
 class DistributedSparseGridDataContainer {
  public:
@@ -30,10 +35,12 @@ class DistributedSparseGridDataContainer {
     kahanDataBegin_.resize(dsgu_.getNumSubspaces(), nullptr);
   }
 
+#ifndef DOXYGEN_SHOULD_SKIP_THIS
   DistributedSparseGridDataContainer(const DistributedSparseGridDataContainer&) = delete;
   DistributedSparseGridDataContainer& operator=(const DistributedSparseGridDataContainer&) = delete;
   DistributedSparseGridDataContainer(DistributedSparseGridDataContainer&&) = default;
   DistributedSparseGridDataContainer& operator=(DistributedSparseGridDataContainer&&) = default;
+#endif  // DOXYGEN_SHOULD_SKIP_THIS
 
   // allocates memory for subspace data and sets pointers to subspaces
   void createSubspaceData() {
@@ -146,7 +153,10 @@ class DistributedSparseGridDataContainer {
   std::vector<FG_ELEMENT> kahanData_;  // Kahan summation residual terms
 };
 
-/* This class can store a distributed sparse grid with a uniform space
+/**
+ * @brief a distributed sparse grid with uniform domain decomposition (on all process groups)
+ *
+ * This class can store a distributed sparse grid with a uniform space
  * decomposition. During construction no data is created and the data size of
  * the subspaces is initialized to zero (data sizes are usually set the
  * first time by registering the dsg in a distributed fullgrid during local
@@ -162,112 +172,224 @@ class DistributedSparseGridUniform : public AnyDistributedSparseGrid {
   // should be enough for the current scenario (cf. test_createTruncatedHierarchicalLevels_large)
   // using SubspaceIndexType = AnyDistributedSparseGrid::SubspaceIndexType;
 
-  /** create sparse grid of dimension d and specify for each dimension the
-   * maximum discretization level
+  /**
+   * @brief create sparse grid
+   *
    * No data is allocated and the sizes of the subspace data is initialized to 0.
+   *
+   * @param dim dimensionality of the grid
+   * @param lmax maximum level of the grid
+   * @param lmin minimum level of the grid
+   * @param comm communicator for the grid
    */
   explicit DistributedSparseGridUniform(DimType dim, const LevelVector& lmax,
                                         const LevelVector& lmin, CommunicatorType comm);
 
   /**
-   * create an empty (no data) sparse grid with given subspaces.
+   * @brief create an empty (no data) sparse grid with given subspaces.
    */
[…]
