### 2024-10-28 4b10d4d
docs: document and reorder CombiCom functions
