+++ src/combicom/CombiCom.hpp
@@ -12,43 +12,18 @@ namespace combigrid {
 
 namespace CombiCom {
 
-template <typename FG_ELEMENT>
-void FGReduce(FullGrid<FG_ELEMENT>& fg, RankType r, MPI_Comm comm) {
-  if (!fg.isGridCreated()) fg.createFullGrid();
-
-  auto& sendbuf = fg.getElementVector();
-  std::vector<FG_ELEMENT> recvbuf(sendbuf.size());
-
-  int myrank;
-  MPI_Comm_rank(comm, &myrank);
-
-  MPI_Datatype dtype =
-      abstraction::getMPIDatatype(abstraction::getabstractionDataType<FG_ELEMENT>());
-
-  if (myrank == r) {
-    MPI_Reduce(MPI_IN_PLACE, &sendbuf[0], static_cast<int>(sendbuf.size()), dtype, MPI_SUM, r,
-               comm);
-  } else {
-    MPI_Reduce(&sendbuf[0], &recvbuf[0], static_cast<int>(sendbuf.size()), dtype, MPI_SUM, r, comm);
-  }
-}
-
-template <typename FG_ELEMENT>
-void FGAllreduce(FullGrid<FG_ELEMENT>& fg, MPI_Comm comm) {
-  if (!fg.isGridCreated()) fg.createFullGrid();
-
-  auto& sendbuf = fg.getElementVector();
-  std::vector<FG_ELEMENT> recvbuf(sendbuf.size());
-
-  MPI_Datatype dtype =
-      abstraction::getMPIDatatype(abstraction::getabstractionDataType<FG_ELEMENT>());
-
-  int myrank;
-  MPI_Comm_rank(comm, &myrank);
-
-  MPI_Allreduce(MPI_IN_PLACE, &sendbuf[0], static_cast<int>(sendbuf.size()), dtype, MPI_SUM, comm);
-}
-
+/**
+ * @brief check if the allocated subspace sizes of a DistributedSparseGrid are correct
+ *
+ * They must be either 0 or the same as the declared size of the subspace.
+ * Only fails if the sizes are different and the program was compiled in Debug mode;
+ * otherwise only prints to std::cout.
+ *
+ * @tparam SparseGridType (derived from) DistributedSparseGridUniform
+ * @param dsg the DistributedSparseGrid
+ * @param subspaceSizes the declared sizes of the allocated subspaces
+ * @return the sum of all elements in \p subspaceSizes
+ */
 template <typename SparseGridType>
 size_t sumAndCheckSubspaceSizes(const SparseGridType& dsg,
                                 const std::vector<SubspaceSizeType>& subspaceSizes) {
@@ -62,7 +37,7 @@ size_t sumAndCheckSubspaceSizes(const SparseGridType& dsg,
     if (!check) {
       int rank;
       MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-      std::cout << "l = " << dsg.getLevelVector(i) << " "
+      std::cout << "sumAndCheckSubspaceSizes; l = " << dsg.getLevelVector(i) << " "
                 << "rank = " << rank << " "
                 << "ssize = " << subspaceSizes[i] << " "
                 << "dsize = " << dsg.getDataSize(i) << std::endl;
@@ -74,16 +49,39 @@ size_t sumAndCheckSubspaceSizes(const SparseGridType& dsg,
   return bsize;
 }
 
+/**
+ * @brief check if all allocated subspace sizes of a DistributedSparseGrid are correct
+ *
+ * like sumAndCheckSubspaceSizes for all subspace data sizes of the DistributedSparseGrid
+ *
+ * @tparam SparseGridType (derived from) DistributedSparseGridUniform
+ * @param dsg the DistributedSparseGrid
[…]
