### 2024-10-29 586310b
docs: document and format CombiParameters
