+++ src/manager/CombiParameters.hpp
@@ -11,10 +11,49 @@
 #include "utils/Types.hpp"
 namespace combigrid {
 
+/**
+ * @brief Class for the parameters of the combination technique
+ *
+ * This class contains all the parameters that describe a combination technique setup.
+ * The class is serializable and can be sent over MPI.
+ * It is used, for instance, by ProcessManager to distribute the combination technique information
+ * to the ProcessGroupWorkers.
+ */
 class CombiParameters {
  public:
-  CombiParameters() {}
+  /**
+   * @brief default constructor for serialization
+   */
+  CombiParameters() = default;
 
+  /**
+   * @brief Constructor for CombiParameters
+   *
+   * the constructor variant with levels, coeffs and taskIDs specified, such that Tasks can be
+   * created and tracked suitable for manager-worker setups
+   *
+   * @param dim dimensionality of the PDE solver domain
+   * @param lmin minimum level vector, size \p dim
+   * @param lmax maximum level vector, size \p dim
+   * @param boundary boundary conditions, size \p dim
+   * @param levels level vectors for each task
+   * @param coeffs coefficients for each task
+   * @param taskIDs IDs for each task
+   * @param numberOfCombinations total number of combinations, only used if
+   * reduceCombinationDimsLmax is used
+   * @param numGrids number of grids per task, \see Task::getDistributedFullGrid()
+   * @param combinationVariant variant of the combination reduction
+   * @param parallelization number of processes in (Cartesian) communicator
+   * @param reduceCombinationDimsLmin unused
+   * @param reduceCombinationDimsLmax reduce the maximum level vector for the distributed sparse
+   * grid (to avoid unnecessary storage / communication use (1, 1, 1, ...))
+   * @param sizeForChunkedCommunicationInMebibyte size of the chunks for chunked reductions in MiB
+   * @param forwardDecomposition in case of ambiguity, decompose the full grid such that the middle
+   * grid points are on the "lower" process
+   * @param thirdLevelHost hostname of the third level manager (deprecated)
+   * @param thirdLevelPort port to connect to third level manager (deprecated)
+   * @param thirdLevelPG process group that should connect to the third level manager (deprecated)
+   */
   CombiParameters(DimType dim, const LevelVector& lmin, const LevelVector& lmax,
                   const std::vector<BoundaryType>& boundary, const std::vector<LevelVector>& levels,
                   const std::vector<real>& coeffs, const std::vector<size_t>& taskIDs,
@@ -54,8 +93,31 @@ class CombiParameters {
     }
   }
 
-  // constructor variant w/o combination scheme specified -- the workers have their partial list
-  // imlpicitly as tasks vector
+  /**
+   * @brief Constructor for CombiParameters
+   *
+   * constructor variant w/o combination scheme specified -- the workers have their partial list
+   * imlpicitly as tasks vector, e.g. worker-only setups
+   *
+   * @param dim dimensionality of the PDE solver domain
+   * @param lmin minimum level vector, size \p dim
+   * @param lmax maximum level vector, size \p dim
+   * @param boundary boundary conditions, size \p dim
+   * @param numberOfCombinations total number of combinations, only used if
+   * reduceCombinationDimsLmax is used
+   * @param numGrids number of grids per task, \see Task::getDistributedFullGrid()
+   * @param combinationVariant variant of the combination reduction
+   * @param parallelization number of processes in (Cartesian) communicator
[…]
