### 2024-10-30 27c4618
docs: document and format ProcessManager
