### 2024-10-31 38e8396
Fix selalib build error (#138)
* selalib: update combine_selalib_phi to current DFG interface

* jenkins: see if discotec +selalib can build

* jenkins: no selalib, revert b1dc7c89e for now

---------

Co-authored-by: Theresa Pollinger <theresa.pollinger@riken.jp>
