### 2024-10-31 ef1f70b
docs: document and format ProcessGroupManager
