+++ src/manager/ProcessGroupManager.hpp
@@ -16,88 +16,224 @@
 #include "utils/Types.hpp"
 
 namespace combigrid {
+/**
+ * @class ProcessGroupManager
+ * @brief The ProcessGroupManager is part of a ProcessManager and is responsible for communication
+ * with a single process group.
+ *
+ * Through the use of non-blocking operations in the ProcessGroupManager, the ProcessManager can
+ * instruct multiple process groups at once.
+ */
 class ProcessGroupManager {
- public:
-  ProcessGroupManager() = delete;
+  friend class ProcessManager;
 
+ public:
+  /**
+   * Constructor
+   *
+   * @param pgroupRootID the rank of each process of this group in the global communicator;
+   * this is the same as the process group number, as the global communicator contains one member of
+   * each process group
+   */
   explicit ProcessGroupManager(RankType pgroupRootID);
 
+#ifndef DOXYGEN_SHOULD_SKIP_THIS
+  ProcessGroupManager() = delete;
+  ~ProcessGroupManager() = default;
+  // rule of 5
   ProcessGroupManager(ProcessGroupManager const&) = delete;
-
   ProcessGroupManager& operator=(ProcessGroupManager const&) = delete;
+  ProcessGroupManager(ProcessGroupManager&&) = delete;
+  ProcessGroupManager& operator=(ProcessGroupManager&&) = delete;
+#endif  // DOXYGEN_SHOULD_SKIP_THIS
 
+  /**
+   * @brief signal the process group to initalize and run the task \p t
+   */
   bool runfirst(Task* t);
 
+  /**
+   * @brief signal the process group to run a time step on all of its tasks
+   */
   bool runnext();
 
+  /**
+   * @brief signal the process group to initialized its sparse grid data structures
+   */
   bool initDsgus();
 
+  /**
+   * @brief signal the process group to exit
+   */
   bool exit();
 
-  /* non-blocking call to retrieve status of process group */
+  /**
+   * @brief non-blocking call to retrieve status of process group
+   */
   inline StatusType getStatus();
 
-  inline void setStatus(StatusType status);
-
-  /* blocks until process group finished computation */
+  /**
+   * @brief blocks until process group finished computation
+   */
   inline StatusType waitStatus();
 
+  /**
+   * @brief get a collection of all tasks currently assigned to this group
+   */
   inline const TaskContainer& getTaskContainer() const;
 
+  /**
+   * @brief remove a task from the process group
+   *
+   * this does not change the state in the workers!
+   */
   inline void removeTask(Task* t);
 
+  /**
+   * @brief signal to perform a system-wide combination
+   */
   bool combine();
 
-  // third Level stuff
+  /**
+   * @brief signal to perform a widely-distributed combination
+   *
+   * based on TCP/socket setup with third level manager. cf. ProcessManager::combineThirdLevel
+   */
   bool combineThirdLevel(const ThirdLevelUtils& thirdLevel, CombiParameters& params,
                          bool isSendingFirst);
-
+  /**
+   * @brief signal to perform a whole widely-distributed combination
+   *
+   * based on file-exchange mechanism (w/o third level manager); equivalent to calling
+   * combineThirdLevelFileBasedWrite and combineThirdLevelFileBasedReadReduce in succession
+   */
   bool combineThirdLevelFileBased(const std::string& filenamePrefixToWrite,
                                   const std::string& writeCompleteTokenFileName,
                                   const std::string& filenamePrefixToRead,
                                   const std::string& startReadingTokenFileName);
 
+  /**
+   * @brief signal to start a widely-distributed combination
+   *
+   * based on file-exchange mechanism (w/o third level manager)
+   */
   bool combineThirdLevelFileBasedWrite(const std::string& filenamePrefixToWrite,
[…]
