### 2024-10-31 370bdfc
docs: document and format ProcessGroupWorker
