+++ src/manager/ProcessGroupWorker.hpp
@@ -10,127 +10,305 @@
 
 namespace combigrid {
 
+/**
+ * @class ProcessGroupWorker
+ *
+ * @brief The ProcessGroupWorker comprises a TaskWorker and a SparseGridWorker;
+ * it is responsible for executing the time stepping in the combination technique as well as various
+ * I-O operations.
+ *
+ * When using a manager-worker scheme, the ProcessGroupWorker is instantiated on each worker rank;
+ * else, it is instantiated on each rank.
+ */
 class ProcessGroupWorker {
  public:
   explicit ProcessGroupWorker();
 
+#ifndef DOXYGEN_SHOULD_SKIP_THIS
   ProcessGroupWorker(ProcessGroupWorker const&) = delete;
   ProcessGroupWorker& operator=(ProcessGroupWorker const&) = delete;
+  ProcessGroupWorker(ProcessGroupWorker&&) = delete;
+  ProcessGroupWorker& operator=(ProcessGroupWorker&&) = delete;
+  ~ProcessGroupWorker() = default;
+#endif  // DOXYGEN_SHOULD_SKIP_THIS
 
-  ~ProcessGroupWorker();
-
-  /** wait for command from manager */
+  /** @brief wait for command from manager */
   SignalType wait();
 
+  /**
+   * @brief call run on each Task once
+   */
   void runAllTasks();
 
+  /**
+   * @brief release all tasks
+   */
   void exit();
 
+  /**
+   * @brief get the vector of tasks
+   */
   inline const std::vector<std::unique_ptr<Task>>& getTasks() const;
 
-  /** initializes all subspace sizes in the dsgu according to the dfgs in the
-   * global reduce comm*/
+  /**@brief  initializes all subspace sizes in the distributed sparse grid data structure */
   void initCombinedDSGVector();
 
-  /** extracts and dehierarchizes */
+  /**
+   * @brief extracts data from sparse to component grids and dehierarchizes the component grids
+   */
   void updateFullFromCombinedSparseGrids();
 
+  /**
+   * @brief dehierarchize all component grids
+   */
   void dehierarchizeAllTasks();
 
-  /** combine on sparse grid with uniform decomposition of domain */
+  /**
+   * @brief do a whole combination, assuming a single-system setup
+   *
+   * @param collectMinMaxCoefficients if true, the min and max coefficients per subspace will be
+   * collected as part of the combination
+   */
   void combineAtOnce(bool collectMinMaxCoefficients = false);
 
+  /**
+   * @brief perform the first part of a system-wide combination,  namely hierarchization
+   * and reduction (but not dehierarchization)
+   *
+   * based on file-exchange mechanism (w/o third level manager)
+   */
   void combineSystemWide();
 
+  /**
+   * @brief perform the first part of a system-wide combination, but reducing to output
+   * ranks only instead of all ranks, which then write data to file.
+   *
+   * based on file-exchange mechanism (w/o third level manager)
+   */
   void combineSystemWideAndWrite(const std::string& writeSparseGridFile,
                                  const std::string& writeSparseGridFileToken);
 
-  /** parallel file io of final output grid */
+  /**
+   * @brief interpolate the current solution from all component grids on this group
+   *
+   * receives the target level and file name from the ProcessGroupManager
+   */
   void parallelEval();
 
-  /** parallel file io of final output grid for uniform decomposition */
+  /**
+   * @brief called by parallelEval to interpolate at resolution level \p leval and write the results
+   * to a binary file readable with Paraview
+   */
   void parallelEvalUniform(const std::string& filename, const LevelVector& leval) const;
 
-  // do task-specific postprocessing
+  /** @brief do task-specific diagnostics / postprocessing */
[…]
