### 2024-11-01 b6241c2
tests: set longer timeouts to avoid CI failures
