+++ tests/test_thirdLevel.cpp
@@ -860,7 +860,7 @@ void testPretendThirdLevel(TestParams& testParams) {
 }
 
 BOOST_FIXTURE_TEST_SUITE(thirdLevel, TestHelper::BarrierAtEnd,
-                         *boost::unit_test::timeout(2500) * boost::unit_test::disabled())
+                         *boost::unit_test::timeout(3500) * boost::unit_test::disabled())
 
 BOOST_AUTO_TEST_CASE(test_0, *boost::unit_test::tolerance(TestHelper::tolerance) *
                                  boost::unit_test::disabled()) {
@@ -1143,7 +1143,7 @@ BOOST_AUTO_TEST_CASE(test_workers_2d, *boost::unit_test::tolerance(TestHelper::t
 
 // same as test_8 but only with workers
 BOOST_AUTO_TEST_CASE(test_8_workers, *boost::unit_test::tolerance(TestHelper::tolerance) *
-                                         boost::unit_test::timeout(1250)) {
+                                         boost::unit_test::timeout(2250)) {
   unsigned int numSystems = 2;
   unsigned int nprocs = 1;
   unsigned int ncombi = 3;
