+++ tests/test_distributedsparsegrid.cpp
@@ -297,7 +297,7 @@ void checkDistributedSparsegrid(LevelVector& lmin, LevelVector& lmax, std::vecto
   }
 }
 
-BOOST_AUTO_TEST_SUITE(distributedsparsegrid, *boost::unit_test::timeout(1800))
+BOOST_AUTO_TEST_SUITE(distributedsparsegrid, *boost::unit_test::timeout(2500))
 // very cheap
 BOOST_AUTO_TEST_CASE(test_0) {
   LevelVector lmin = {1, 1};
