### 2024-11-01 e0e0aba
docs: document SparseGridWorker
