+++ src/manager/SparseGridWorker.hpp
@@ -12,108 +12,320 @@
 #include "sparsegrid/DistributedSparseGridUniform.hpp"
 namespace combigrid {
 
-// template <typename CombiDataType, DimType NumDimensions>
+/**
+ * @class SparseGridWorker
+ *
+ * @brief The SparseGridWorker class is responsible for managing the distributed sparse grids
+ */
 class SparseGridWorker {
+  // template <typename CombiDataType, DimType NumDimensions>
  public:
+  /**
+   * @brief Constructor
+   *
+   * @param taskWorkerToReference the task worker that this sparse grid worker should reference
+   */
   explicit SparseGridWorker(TaskWorker& taskWorkerToReference);
 
+#ifndef DOXYGEN_SHOULD_SKIP_THIS
   SparseGridWorker(SparseGridWorker const&) = delete;
   SparseGridWorker& operator=(SparseGridWorker const&) = delete;
 
   ~SparseGridWorker() = default;
+#endif  // DOXYGEN_SHOULD_SKIP_THIS
 
+  /**
+   * @brief collect, reduce, and distribute the sparse grid data
+   *
+   * @tparam keepValuesInExtraSparseGrid if true, the values that would be useful to the extra
+   * sparse grid are also gathered there (necessary for widely-distributed chunked reduction)
+   * @param combinationVariant the combination variant to use
+   * @param maxMiBToSendPerThread the maximum amount of data to send per OpenMP thread at once
+   * @param collectMinMaxCoefficients if true, the min and max hierarchical coefficients are
+   * collected
+   */
   template <bool keepValuesInExtraSparseGrid = false>
   inline void collectReduceDistribute(CombinationVariant combinationVariant,
                                       uint32_t maxMiBToSendPerThread,
                                       bool collectMinMaxCoefficients = false);
 
+  /**
+   * @brief copy data from the extra sparse grid to the partially allocated combined sparse grid
+   *
+   * necessary for widely-distributed chunked reduction; copies only the subspaces allocated in both
+   */
   inline void copyFromExtraDsgToPartialDSG(size_t gridNumber = 0);
 
+  /**
+   * @brief copy data from the partially allocated combined sparse grid to the extra sparse grid
+   *
+   * copies only the subspaces allocated in both.
+   *
+   * necessary for widely-distributed chunked reduction; ; only implemented for file-based setup
+   */
   inline void copyFromPartialDsgToExtraDSG(size_t gridNumber = 0);
 
-  /* free DSG memory as intermediate step */
+  /**
+   * @brief free sparse grid data memory
+   */
   inline void deleteDsgsData();
 
+  /**
+   * @brief chunked broadcast of the sparse grid data
+   *
+   * necessary for widely-distributed chunked reduction; only implemented for file-based setup
+   *
+   * @param maxMiBToSendPerThread the maximum amount of data to send per OpenMP thread at once
+   */
   inline void distributeChunkedBroadcasts(uint32_t maxMiBToSendPerThread);
 
+  /**
+   * @brief update the tasks' full grids from the hierarchical sparse grid data
+   */
   inline void distributeCombinedSolutionToTasks();
 
+  /**
+   * @brief fill a task's full grid from the distributed sparse grid and dehierarchize
+   *
+   * can only be used for non-chunked reductions
+   *
+   * @param dfg the distributed full grid to fill
+   * @param boundary the boundary conditions used for hierarchization
+   * @param hierarchizationDims the hierarchized dimensions
+   * @param hierarchicalBases the hierarchical bases used for hierarchization
+   * @param lmin the minimum level vector the grids are hierarchized to
+   */
   inline void fillDFGFromDSGU(Task& t, const std::vector<bool>& hierarchizationDims,
[…]
