### 2024-11-01 9f5ec12
docs: document and format MPISystem,  and some files that touch it
