+++ src/mpi/MPISystem.hpp
@@ -13,7 +13,7 @@
 #include "mpi_fault_simulator/MPI-FT.h"
 #include "utils/Types.hpp"
 
-#define MASTER_EXCLUSIVE_SECTION if (combigrid::theMPISystem()->isMaster())
+#define MASTER_EXCLUSIVE_SECTION if (combigrid::theMPISystem()->isProcessGroupMaster())
 
 // first group
 #define FIRST_GROUP_EXCLUSIVE_SECTION if (combigrid::theMPISystem()->getProcessGroupNumber() == 0)
@@ -42,12 +42,16 @@
           (combigrid::theMPISystem()->getNumGroups() * combigrid::theMPISystem()->getNumProcs()) / \
           2))
 
-#define GLOBAL_MANAGER_EXCLUSIVE_SECTION if (combigrid::theMPISystem()->isGlobalManager())
-
 #define WORLD_MANAGER_EXCLUSIVE_SECTION if (combigrid::theMPISystem()->isWorldManager())
 
 namespace combigrid {
 
+/**
+ * @brief RAII class to turn MPI on and off
+ *
+ * This class initialized MPI at creation and finalizes at destruction (e.g. when the object goes
+ * out of scope).
+ */
 struct [[nodiscard]] MpiOnOff {
   explicit MpiOnOff(int* argc = nullptr, char*** argv = nullptr);
   ~MpiOnOff();
@@ -57,7 +61,8 @@ class ProcessGroupManager;
 
 class MPISystem;
 typedef std::shared_ptr<MPISystem> MPISystemID;
-/** MPI Communication System
+/**
+ * @class MPISystem
  *
  * This class encapsulates the access to the different communicators being used.
  *
@@ -65,94 +70,86 @@ typedef std::shared_ptr<MPISystem> MPISystemID;
  * process does not belong to a specific communicator the corresponding get function
  * returns MPI_COMM_NULL.
  *
- * CommWorld: contains all processes. usually equal to MPI_COMM_WORLD. if fault
+ * worldComm_: contains all processes. usually equal to MPI_COMM_WORLD. if fault
  * tolerance is enabled the processes of groups detected as failed will be removed
  *
- * GlobalComm: contains the manager and the master process of each process group
+ * globalComm_: contains the manager and the master process of each process group
  *
- * SpareCommFT: contains inactive but reusable ranks. Only used for FT simulations.
+ * spareCommFT_: contains inactive but reusable ranks. Only used for fault-tolerance
+ * simulations.
  *
- * LocalComm: contains the processes of the group a process is assigned to. this
+ * localComm_: contains the processes of the group a process is assigned to. this
  * is MPI_COMM_NULL for the manager process. this is the communicator the
- * application MUST use for the computations. the application MUST NEVER use
- * MPI_COMM_WORLD internally.
+ * Task uses for the computations.
  *
- * GlobalReduceComm: contains the processes for the global reduction. in the
+ * globalReduceComm_: contains the processes for the global reduction. in the
  * case of equally sized process groups (at the moment this is the only option)
  * this communicator contains all processes which have the same rank in
- * LocalComm. it is MPI_COMM_NULL on the manager.
+ * localComm_. it is MPI_COMM_NULL on the manager.
  *
- * ThirdLevelComms: list of communicators for the thirdLevelCombination. Each
+ * thirdLevelComms_: list of communicators for the thirdLevelCombination. Each
  * communicator connects all workers of a pg to
  * the process manager. The list differs for each caller and contains only the
  * comms where he participates: The process manager gets all comms whereas each
  * worker has a single entry.
  *
- * getXXXCommFT returns the fault tolerant equivalent of Communicator XXX
- *
[…]
