### 2024-11-01 SGpp/DisCoTec#135 @freifrauvonbleifrei
I increased the timeouts (in https://github.com/SGpp/DisCoTec/commit/b6241c2237e6e206b9ba13714bcd0eab5897b01c) and default-disabled the file-based tests. thanks for pointing it out!
