### 2024-11-01 #7018 @freifrauvonbleifrei
Sorry, I had forgotten to push my added documentation to here: https://github.com/SGpp/DisCoTec/pull/133

I am currently trying to remove the worst documentation clutter and am on track to merge today :)
