### 2024-11-01 07bab7c
CombiParameters: default value for value member variable
