### 2024-11-01 232946b
Add Readthedocs Documentation for DisCoTec (#133)
* docs: add basic doxygen and sphinx configuration and readthedocs config
* docs: add Doxygen markup to most important classes
* docs: add some docs in .rst format, some adapted from READMEs
* docs: rework some READMEs
* remove some unused functions and includes
* adds Code of Conduct and Contributing guideline
* tests: have slightly longer timeouts

---------

Co-authored-by: Theresa Pollinger <theresa.pollinger@riken.jp>
Co-authored-by: Michael Obersteiner <oberstei@in.tum.de>
Co-authored-by: Alexander Van Craen <40516079+vancraar@users.norep
[…]
