+++ src/manager/ProcessManager.hpp
@@ -1,182 +1,387 @@
 #ifndef PROCESSMANAGER_HPP_
 #define PROCESSMANAGER_HPP_
 
-#include <vector>
 #include <numeric>
+#include <vector>
 
 #include "combischeme/CombiMinMaxScheme.hpp"
 #include "fault_tolerance/LPOptimizationInterpolation.hpp"
+#include "loadmodel/LearningLoadModel.hpp"
+#include "loadmodel/LoadModel.hpp"
 #include "manager/ProcessGroupManager.hpp"
 #include "manager/ProcessGroupSignals.hpp"
-#include "loadmodel/LoadModel.hpp"
-#include "loadmodel/LearningLoadModel.hpp"
-#include "rescheduler/TaskRescheduler.hpp"
-#include "rescheduler/StaticTaskRescheduler.hpp"
 #include "mpi/MPISystem.hpp"
+#include "rescheduler/StaticTaskRescheduler.hpp"
+#include "rescheduler/TaskRescheduler.hpp"
 #include "task/Task.hpp"
-#include "manager/ProcessGroupManager.hpp"
-#include "combischeme/CombiMinMaxScheme.hpp"
 #include "third_level/ThirdLevelUtils.hpp"
 #include "utils/MonteCarlo.hpp"
 
 namespace combigrid {
 
+/**
+ * @brief The ProcessManager class orchestrates the whole simulation in case of a manager-worker
+ * scheme in DisCoTec.
+ *
+ * It should only be instantiated once---on the manager rank---and it holds one ProcessGroupManager
+ * instance for communication with each of the process groups.
+ *
+ * All other ranks should instantiate a ProcessGroupWorker and call wait() on the worker, until an
+ * exit signal is received.
+ */
 class ProcessManager {
  public:
-   /**
-    * Constructor for a process manager.
-    *
-    * @param pgroups The process groups.
-    * @param instances The tasks.
-    * @param params The parameters for the combination technique.
-    * @param loadModel The load model to use for scheduling. If it is a 
-    *                  learning load model duration information is added for 
-    *                  every task after every run.
-    * @param rescheduler The rescheduler to use for dynamic task rescheduling.
-    *                    By default the static task rescheduler is used and 
-    *                    therefore no rescheduling perfomed.
-    */
+  /**
+   * @brief Constructor
+   *
+   * @param pgroups a vector of ProcessGroupManager s
+   * @param instances a vector of Task s
+   * @param params the combination parameters
+   * @param loadModel The load model to use for (re)scheduling
+   * @param rescheduler The rescheduler to use for dynamic task rescheduling.
+   *                    By default, the static task rescheduler is used and
+   *                    therefore no rescheduling perfomed.
+   */
   ProcessManager(ProcessGroupManagerContainer& pgroups, TaskContainer& instances,
-                 CombiParameters& params, std::unique_ptr<LoadModel> loadModel, 
-                 std::unique_ptr<TaskRescheduler> rescheduler = std::unique_ptr<TaskRescheduler>(new StaticTaskRescheduler{}))
-    : pgroups_{pgroups},
-      tasks_{instances},
-      params_{params},
-      loadModel_{std::move(loadModel)},
-      rescheduler_{std::move(rescheduler)},
-      thirdLevel_{params.getThirdLevelHost(), params.getThirdLevelPort()},
-      thirdLevelPGroup_{pgroups_[params.getThirdLevelPG()]}
-  {
-     // only setup third level if explicitly desired
-     if (params.getThirdLevelHost() != "")
-       setupThirdLevel();
+                 CombiParameters& params, std::unique_ptr<LoadModel> loadModel,
[…]
