### 2024-11-10 9b5f190
docs: "PDE solver" in all markdown files
