### 2024-11-11 8cac08a
docs: address first two comments of https://github.com/SGpp/DisCoTec/issues/129#issuecomment-2383274769
