### 2024-11-13 #7018 @freifrauvonbleifrei
@danielskatz good question!

I was hoping that now @EmilyBourne and @jakelangham would start their hands-ons, while I further refine the documentation (progress here https://github.com/SGpp/DisCoTec/tree/refine_documentation) according to the things that were already mentioned here: https://github.com/SGpp/DisCoTec/issues/129 , and trying to make the Doxygen output more usable.

probably issue 129 would also be a good spot for further docs-related feedback to give things to incorporate on a rolling basis ;)
