### 2024-11-24 #7018 @danielskatz
👋 @EmilyBourne & @jakelangham - just checking with you on [this](https://github.com/openjournals/joss-reviews/issues/7018#issuecomment-2472864290) again...
