### 2024-12-05 #7018 @jakelangham
Hi, @danielskatz, have I read back through the thread and hope there hasn't been a miscommunication here. I finished my review (in the sense of interrogating every bullet point) some time ago and was waiting for the points I raised to be addressed, after which I expected to be happy to recommend for publication.

Specifically, I can see https://github.com/SGpp/DisCoTec/issues/134, https://github.com/SGpp/DisCoTec/issues/137, https://github.com/SGpp/DisCoTec/issues/129. As far as I'm aware there has been some progress but they're still being worked on.
