### 2024-12-05 #7018 @danielskatz
Thanks @jakelangham. It's helpful to have this in one place, and I apologize if it seemed like I was blaming you; I was just trying to pull all the statuses together.

👋 @freifrauvonbleifrei - can you update us here on how the work on these open issues are going?
