### 2024-12-06 abd06b8
docs: first draft for issue #134
