### 2024-12-06 #7018 @freifrauvonbleifrei
@jakelangham thx for the overview! I had started working diffusely on the different questions but it really helps to see them as a (finite) list! :D
some feedback requests follow in the respective issues.
