### 2024-12-06 SGpp/DisCoTec#134 @freifrauvonbleifrei
@jakelangham this is my first try -- I tried to write it in a way that would not introduce a lot of symbols that are not already present in the image, and to mention only things that are so general I would not have to point to one specific source.

Please let me know what you think! I feel like anything Sparse Grid quickly generates a whole tail of notation, so I am rather unsure if this is the right "depth". If you agree, I would put a similar range of additional info into the paper (or maybe only the complexities?).
