### 2024-12-06 SGpp/DisCoTec#134 @freifrauvonbleifrei
(see here for the actual rendering -> https://discotec--139.org.readthedocs.build/en/139/combination_technique.html )
