### 2024-12-06 SGpp/DisCoTec#137 @freifrauvonbleifrei
> This doesn't necessarily need to go in the paper, but could you explain how using OpenMP decreases the memory footprint?
Sure: At the core counts we are targeting, it gets impractical to use one MPI rank for every core. This is because each rank needs to store some information about addressing every other rank. Trading MPI ranks for OpenMP threads helps here.
Probably the reference on saving memory can go? I was still quite proud about it when we worked on the paper draft, but I think your question shows that it brings more confusion than clarity.
