### 2024-12-06 af7f898
docs: minor edits to combination_technique
