### 2024-12-06 SGpp/DisCoTec#129 @freifrauvonbleifrei
> In the second figure of 'Paralelism in DisCoTec', it might be nice to indicate the underlying problem that was being solved up front. You refer to 'the advection solver step' but that context needs to be set up in advance.

I feel some inner resistance to actually introduce the solver here, because I think it would distract from the main point I am trying to make. What is your opinion on replacing "advection solver" with "memory-bandwidth-bound 6-D PDE solver" (or similar) in this section @jakelangham ?
