### 2024-12-06 f4bc43d
docs: expand process group discussion
