+++ docs/parallelism.md
@@ -16,9 +16,17 @@ Figure originally published in (Pollinger [2024](https://elib.uni-stuttgart.de/h
 
 All component grids are distributed to process groups (either statically, or
 dynamically through the manager rank).
+Each process group will be responsible for a set of component grids and the
+solver instance belonging to each.
+Within a process group and each grid, one can observe the exact same domain 
+decomposition parallelism as when using the solver directly.
+(The domain decomposition needs to be the same for all solver instances.)
+
+During a time step, the PDE solver step applies one or multiple time step updates
+to the values in each component grid.
 During the PDE solver time step and most of the combination step, MPI communication
 only happens within the process groups.
-Conversely, for the sparse grid reduction using the combination coefficients,
+Conversely, for the sparse grid reduction using the combination coefficients $c_\vec{\ell}^c$,
 MPI communication only happens between a rank and its colleagues in the other
 process groups, e.g., rank 0 in group 0 will only talk to rank 0 in all other groups.
 Thus, major bottlenecks arising from global communication can be avoided altogether.
@@ -44,7 +52,7 @@ parallelizations](../gfx/times-solver-on-hawk.svg)
 ![timings for combination step on
 HAWK at various parallelizations](../gfx/times-combination-on-hawk.svg)
 
-We see the timings (in seconds) for the advection solver step and the
+We see the timings (in seconds) for the advection PDE solver step and the
 combination step, respectively.
 This weak scaling experiment used four OpenMP threads per rank, and starts with
 one pg of four processes in the upper left corner.
@@ -52,8 +60,12 @@ The largest parallelization is 64 pgs of 2048 processes each.
 Figure originally published in (Pollinger [2024](https://elib.uni-stuttgart.de/handle/11682/14229)).
 
 This image also describes the challenges in large-scale experiments with DisCoTec:
-If the process groups become too large, the MPI communication of the multiscale
+Generally, the process group size can be chosen so that it matches the solver
+and the desired resolutions well.
+But if the process groups become too large, the MPI communication of the multiscale
 transform starts to dominate the combination time.
+Conversely, the number of process groups can be freely chosen, as long as there
+is sufficient work for each group to do (one grid per group at a bare minimum).
 If there are too many pgs, the combination reduction will dominate the
 combination time.
 However, the times required for the PDE solver stay relatively constant;
