### 2024-12-06 8a47da2
docs: expand process group discussion (fix formatting, amends last commit)
