### 2024-12-06 85d1576
docs: use 6 dimensions everywhere in tutorial (addresses 6th point raised by @jakelangham in issue #129)
