+++ docs/simple_tutorial.md
@@ -15,6 +15,8 @@ int num_points_x = ...
 int num_points_y = ...
 int num_points_z = ...
 int num_points_vx = ...
+int num_points_vy = ...
+int num_points_vz = ...
 ...
 initial_function = ...
 grid.initialize(
@@ -23,6 +25,8 @@ grid.initialize(
                 num_points_y,
                 num_points_z,
                 num_points_vx,
+                num_points_vy,
+                num_points_vz,
                 ... 
                )
 helper_data_structures.initialize(...)
@@ -38,6 +42,8 @@ while(time_now < end_time) { # time loop
 properties = compute_properties(grid)
 write_output(properties, grid)
 ```
+
+This example assumes that your PDE solver uses three dimensions in both space and velocity.
 
 For use with DisCoTec, we assume nestable power-of-two discretizations, i.e.,
 where your grid spacing can be $2^{-l}$ for $l \in \mathbb{N}$.
@@ -60,6 +66,8 @@ int num_points_x = ...
 int num_points_y = ...
 int num_points_z = ...
 int num_points_vx = ...
+int num_points_vy = ...
+int num_points_vz = ...
 ...
 
 - initial_function = ...
@@ -69,6 +77,8 @@ int num_points_vx = ...
 -                 num_points_y,
 -                 num_points_z,
 -                 num_points_vx,
+-                 num_points_vy,
+-                 num_points_vz,
 -                 ... 
 -                )
 - helper_data_structures.initialize(...)
@@ -77,6 +87,8 @@ int num_points_vx = ...
 +                                num_points_y,
 +                                num_points_z,
 +                                num_points_vx,
++                                num_points_vy,
++                                num_points_vz,
 +                                ...
 +                               )
 float end_time = ...
@@ -112,7 +124,9 @@ C++ header `YourSolver.h`, roughly like this:
 
 class YourSolver {
  public:
-  YourSolver(int num_points_x, int num_points_y, int num_points_vx, int num_points_vy, ...);
+  YourSolver(int num_points_x, int num_points_y, int num_points_z,
+             int num_points_vx, int num_points_vy, int num_points_vz, 
+             ...);
   //...rule of 5...
 
   void run(double time_step);
@@ -143,9 +157,12 @@ With `YourSolver`, that will be as simple as:
 
 class YourTask : public combigrid::Task {
  public:
-  YourTask(combigrid::LevelVector& l, const std::vector<combigrid::BoundaryType>& boundary,
-           combigrid::real coeff, combigrid::real dt)
-      : Task(4, l, boundary, coeff, nullptr, nullptr), sim_(nullptr), dfg_(nullptr), dt_(dt) {}
+  YourTask(combigrid::LevelVector& l, 
+           const std::vector<combigrid::BoundaryType>& boundary,
+           combigrid::real coeff, 
+           combigrid::real dt)
+      : Task(4, l, boundary, coeff, nullptr, nullptr), 
+        sim_(nullptr), dfg_(nullptr), dt_(dt) {}
 
   virtual ~YourTask(){
     if (sim_ != nullptr){
@@ -162,17 +179,20 @@ class YourTask : public combigrid::Task {
     const auto& l = this->getLevelVector();
     int num_points_x = combigrid::powerOfTwoByBitshift(l[0]);
     int num_points_y = combigrid::powerOfTwoByBitshift(l[1]);
-    int num_points_vx = combigrid::powerOfTwoByBitshift(l[2]);
-    int num_points_vy = combigrid::powerOfTwoByBitshift(l[3]);
+    int num_points_z = combigrid::powerOfTwoByBitshift(l[2]);
+    int num_points_vx = combigrid::powerOfTwoByBitshift(l[3]);
+    int num_points_vy = combigrid::powerOfTwoByBitshift(l[4]);
[…]
