### 2024-12-06 7254018
docs: some more urls (addresses last of @jakelangham's points in issue #129 )
