### 2024-12-06 SGpp/DisCoTec#129 @freifrauvonbleifrei
@jakelangham IMHO I have addressed the issues you raised in the current version of PR #139, see here https://discotec--139.org.readthedocs.build/en/139/ (minus the "advection solver", commented above). Let me know in case there is more open for you! (also to @EmilyBourne of course!)
