### 2024-12-06 SGpp/DisCoTec#137 @jakelangham
Interesting. Up to you if you want to keep it in, but if you do, then I think there is room to add a brief explanation without disrupting the flow of the text too much.
