### 2024-12-10 4642d4d
paper: more references, expand related work, define I, software -> references cf. https://github.com/SGpp/DisCoTec/issues/137
