+++ paper.md
@@ -37,17 +37,18 @@ bibliography: paper.bib
 
 # Summary
 
-`DisCoTec` is a C++ framework for the sparse grid combination technique,
+`DisCoTec` is a C++ framework for the sparse grid combination technique [@griebelCombinationTechniqueSolution1992],
 designed for massively parallel settings.
 It is implemented with shared-memory parallelism via OpenMP and
 distributed-memory parallelism via MPI, and is intended to be used in
 conjunction with existing simulation codes.
 For simulation codes that can handle nested structured grids, little to no
 adaptation work is needed for use with the `DisCoTec` framework.
-`DisCoTec` demonstrates its superiority in higher-dimensional time-dependent
-simulations, such as high-fidelity plasma simulations in 4- to 6-dimensions
-[@pollingerStableMassconservingSparse2023] and even for simulations in two
-dimensions, improvements may be observed.
+The combination technique with `DisCoTec` demonstrates its superiority in
+memory-per-precision for higher-dimensional time-dependent simulations, such
+as high-fidelity plasma turbulence simulations
+in four to six dimensions
+and even for simulations in two dimensions, improvements can be observed [@pollingerStableMassconservingSparse2023].
 
 A central part of the combination technique at scale is the transformation of
 grid coefficients into a multi-scale basis.
@@ -58,7 +59,8 @@ from the model order reduction provided by the underlying sparse grid approach
 used by `DisCoTec`, without requiring any multi-scale operations.
 An additional feature of `DisCoTec` is the possibility of performing
 widely-distributed simulations of higher-dimensional problems, where multiple
-HPC systems collaborate to solve a joint simulation, as demonstrated in [@pollingerLeveragingComputePower2023].
+High-Performance Computing (HPC) systems collaborate to solve a joint simulation,
+as demonstrated in [@pollingerRealizingJointExtremeScale2024].
 Thus, `DisCoTec` can leverage the compute power and main memory of multiple HPC
 systems, with comparatively low and manageable transfer costs due to the
 combination technique.
@@ -68,21 +70,28 @@ combination technique.
 Higher-dimensional problems (by which we mean more than three space
 dimensions and one time dimension) quickly require infeasible amounts of
 computational resources such as memory and core-hours as the problem size
-increases---they are haunted by the
-so-called 'curse of dimensionality'.
-An example of this are high-fidelity plasma simulations in the field of confined
-fusion research.
-Current approaches to this problem include dimensionally-reduced models
-(which may not always be applicable), and restricting computations to a very limited resolution.
+increases---they are haunted by the so-called 'curse of dimensionality'.
+An example of this are high-fidelity plasma turbulence simulations in the field
+of confined fusion research.
+Currently employed approaches to this problem include dimensionally-reduced models,
+such as gyrokinetics [@brizardFoundationsNonlinearGyrokinetic2007]
+(which may not always be applicable),
+particle-in-cell methods (which suffer from inherent noise [@verboncoeurParticleSimulationPlasmas2005]),
+and restricting computations to a very limited resolution.
+A further---still developing but very promising---approach to the problem are
+low-rank methods [@einkemmerMassMomentumEnergy2021].
 Multi-scale (hierarchical) methods, such as the sparse grid combination
-technique that `DisCoTec` employs, provide an alternative approach to addressing the curse of dimensionality.
-While some implementations of the sparse grid combination technique are
-available in the context of UQ, there is currently no other implementation for
+technique (CT) that `DisCoTec` employs,
+provide an alternative approach to addressing the curse of dimensionality by
[…]
