+++ paper.bib
@@ -6,6 +6,16 @@
   date = {1992},
   pages = {263--281},
   publisher = {IMACS, Elsevier, North Holland},
+  url = {https://ins.uni-bonn.de/media/public/publication-media/griesiam.ps.gz},
+}
+
+@article{bungartzSparseGrids2004,
+  title = {Sparse {{Grids}}},
+  author = {Bungartz, Hans-Joachim and Griebel, Michael},
+  date = {2004},
+  journaltitle = {Acta Numerica},
+  volume = {13},
+  pages = {147--269},
 }
 
 @phdthesis{heeneMassivelyParallelCombination2018,
@@ -27,18 +37,19 @@
   date        = {2021}
 }
 
-@inproceedings{pollingerLeveragingComputePower2023,
-  title = {Leveraging the {{Compute Power}} of {{Two HPC Systems}} for {{Higher-Dimensional Grid-Based Simulations}} with the {{Widely-Distributed Sparse Grid Combination Technique}}},
-  booktitle = {Proceedings of the {{International Conference}} for {{High Performance Computing}}, {{Networking}}, {{Storage}} and {{Analysis}}},
-  author = {Pollinger, Theresa and Van Craen, Alexander and Niethammer, Christoph and Breyer, Marcel and Pflüger, Dirk},
-  date = {2023-11-11},
-  series = {{{SC}} '23},
-  publisher = {{Association for Computing Machinery}},
-  location = {{New York, NY, USA}},
-  doi = {10.1145/3581784.3607036},
-  url = {https://dl.acm.org/doi/10.1145/3581784.3607036},
-  abstract = {Grid-based simulations of hot fusion plasmas are often severely limited by computational and memory resources; the grids live in four- to six-dimensional space and thus suffer the curse of dimensionality. However, high resolutions are required to fully capture the physics of interest. The sparse grid combination technique is a multi-scale method in which many anisotropically coarse resolved grids are used to approximate a fine-scale solution---and it alleviates the curse of dimensionality. This paper presents the core concepts of the widely-distributed combination technique, which allows us to use the compute power and memory of more than one HPC system for the same simulation. We apply the sparse grid combination technique to a six-dimensional advection problem serving as a proxy for plasma simulations. The full-grid solution approximated by the combination technique would contain ≈ 5 ZB if computed with conventional grid-based methods. Even the combination technique simulation operates on ≈ 1 × 1011 double-precision degrees of freedom, or 988 GB, plus the supporting sparse grid data structures. We propose a new approach to divide the compute load between the two HPC systems, requiring only 76 GB to be exchanged. Based on this, we have realized the first synchronous combination technique simulation using two HPC systems, in our case the two German Tier-0 supercomputers HAWK and SuperMUC-NG. On two systems, the simulation can be computed at an average overhead of ≈ 35 \% (108 s per combination step) for file I/O and transfer. The presented concepts apply to any pair of HPC systems if high-speed data transfer is possible.},
-  keywords = {combination technique,coupling HPC systems,higher-dimensional simulation,multi-level methods,plasma turbulence,UFTP},
+@inproceedings{pollingerRealizingJointExtremeScale2024,
+  title = {Realizing {{Joint Extreme-Scale Simulations}} on {{Multiple Supercomputers}}—{{Two Superfacility Case Studies}}},
+  author = {Pollinger, Theresa and Craen, Alexander Van and Offenhäuser, Philipp and Pflüger, Dirk},
+  date = {2024-11-09},
+  pages = {1568--1584},
[…]
