### 2024-12-10 b41a696
paper: Statement of need/second paragraph -> sentence cf. https://github.com/SGpp/DisCoTec/issues/137
