### 2024-12-10 5bd3bc6
paper: rotate to "precision-per-memory" cf. https://github.com/SGpp/DisCoTec/issues/137
