### 2024-12-10 SGpp/DisCoTec#137 @freifrauvonbleifrei
hi @jakelangham , I decided on a "lightweight" modification on the OpenMP / memory saving topic, but I am not sure about how much easier it really is to get the point:

https://github.com/SGpp/DisCoTec/actions/runs/12250144549/artifacts/2297939731

The second paragraph in "statement of need" however is reduced to one sentence and reads much smoother now :) thx for the good suggestion!
