### 2024-12-10 #7018 @freifrauvonbleifrei
@danielskatz I think that now we are on top of all the issues, but happy to know if there's more we can clarify / extend in the documentation.
(I think I cannot keep exhale from generating the "Class Hierarchy" and "File Hierarchy" in the [Library API](https://discotec--139.org.readthedocs.build/en/139/api/library_root.html), but also I have gotten kind of used to them by now ;) )
