### 2024-12-10 38502c9
paper.bib: add missing DOIs
