+++ paper.bib
@@ -115,6 +115,7 @@
   author = {Brizard, Alain J. and Hahm, Taik Soo},
   date = {2007},
   journaltitle = {Reviews of modern physics},
+  doi = {10.1103/revmodphys.79.421},
   volume = {79},
   number = {2},
   pages = {421},
@@ -166,6 +167,7 @@
   volume = {47},
   pages = {A231},
   publisher = {IOP Publishing},
+  doi = {10.1088/0741-3335/47/5a/017},
   url = {https://iopscience.iop.org/article/10.1088/0741-3335/47/5A/017/meta?casa_token=aElJiSgwIgcAAAAA:xrmEQjmliSWfz_8HcmH7scqGnsFqepsQEqxMBxI-q8uEAvsFo6MSKB46NsoEgCI-oC1k3A3K9z8tyHKEtLsYUptMhe8gWQ},
   urldate = {2024-12-04},
   issue = {5A},
