### 2024-12-16 SGpp/DisCoTec#129 @jakelangham
Hi @freifrauvonbleifrei - thanks for this. I think you really made some great improvements here. I've read through and overall things are much clearer. In particular the considerable work on the API is greatly appreciated. Only 2 tiny comments:

- In 'Parallelism in DisCoTec', there is an `\ell` whose LaTeX is not rendering.
- Looking back at my comment around the 'advection solver' line, I don't think I was asking to define the PDE _per se_. Perhaps it is just a subtle English language thing (or just my subjective feeling), but when you say '_the_ advection PDE solver', it makes me think I must have missed where you told me about this particular PDE previously. So perhaps you could instead say something like  'We see the timings (in seconds) for the solver and combination steps respectively, for an illustrative (6-D) PDE problem.'
