### 2024-12-16 SGpp/DisCoTec#134 @jakelangham
@freifrauvonbleifrei I find this level of detail very helpful, thanks. I think there are a few remaining typos though:

- The combination coefficients are sometimes `c^c` and sometimes `c^_{\vect{\ell}}`. I suppose it should always be the latter.
- `\mathcal{I}` is not defined and I presume the sum definining `f_{SG}` should be over `\vect{\ell} \in \mathcal{I}`.
- In a couple of places the LaTeX is not rendering. There's a `\ell` missing a backslash and also a `\log`.
