### 2024-12-16 SGpp/DisCoTec#137 @jakelangham
Hi @freifrauvonbleifrei - looking good. Only a few minor gripes / typos left to clear up from my perspective:

- As in the new documentation, I think `\ell_{\max}` should actually be `\ell^{\max}` to be consistent with the notation used for elements of these vectors + they need a backslash to render the `max` properly. (This is in the Fig 1 caption.)
- Some of the new references are missing spaces between the last word and the opening `(`.
- On line 76 you don't need to redefine High Performance Computing since it was already taken care of earlier. ;)
- I would suggest consulting the editor on how best to do github citations. These can be a bit hard because of determining attribution etc (though some repositories will have details on how to cite). Maybe there is a preferred way to do this in JOSS.
