### 2024-12-16 #7018 @jakelangham
@freifrauvonbleifrei @danielskatz I've looked through all the improvements and submitted a few suggestions but they are very minor in scope, so I probably don't have more to add here unless there are specific questions about things. In short, I'd be happy if the paper proceeded forward to publication.

@freifrauvonbleifrei Two additional things that need doing beforehand: (1) check the references + DOIs (e.g. link on the last ref, [Verboncouer 2005] might need tidying up somehow); (2) all the improvements need merging into main (excepting the paper, which can live on its own branch, if preferred).
