### 2024-12-17 SGpp/DisCoTec#134 @freifrauvonbleifrei
Hi @jakelangham, thanks for pointing this out! will update! (also, I am starting to think that my editor renders markdown math differently than github)
