### 2024-12-19 1afb710
paper: minor edits cf. https://github.com/SGpp/DisCoTec/issues/137#issuecomment-2545897759
