+++ paper.bib
@@ -96,18 +96,14 @@
   langid = {english}
 }
 
-@online{piazzolaSparseGridsMatlab2022,
-  title = {The {{Sparse Grids Matlab}} Kit -- a {{Matlab}} Implementation of Sparse Grids for High-Dimensional Function Approximation and Uncertainty Quantification},
-  author = {Piazzola, Chiara and Tamellini, Lorenzo},
-  date = {2022-03-17},
-  eprint = {2203.09314},
-  eprinttype = {arXiv},
-  eprintclass = {cs, math},
-  doi = {10.48550/arXiv.2203.09314},
-  url = {http://arxiv.org/abs/2203.09314},
-  urldate = {2023-03-15},
-  abstract = {The Sparse Grids Matlab kit is a collection of Matlab functions for high-dimensional interpolation and quadrature, based on the combination technique form of sparse grids. It is lightweight, high-level and easy to use, good for quick prototyping and teaching. It is somehow geared towards Uncertainty Quantification (UQ), but it is flexible enough for other purposes. The goal of this paper is to give an overview of the implementation structure of the Sparse Grids Matlab kit and to showcase its potentialities, guided by some illustrative tests and a final comprehensive example on forward and inverse UQ analysis.},
-  keywords = {Computer Science - Mathematical Software,Mathematics - Numerical Analysis},
+@article{piazzolaSparseGridsMatlab2024,
+ author = {Piazzola, Chiara and Tamellini, Lorenzo},
+  title  = {{Algorithm 1040: The Sparse Grids Matlab Kit - a Matlab implementation of sparse grids for high-dimensional function approximation and uncertainty quantification}},
+ journal= {ACM Transactions on Mathematical Software},
+ year   = {2024},
+ volume = {50},
+ number = {1},
+ doi = {10.1145/3630023}
 }
 
 @article{brizardFoundationsNonlinearGyrokinetic2007,
@@ -138,14 +134,14 @@
   keywords = {Complexity reduction,Conservative numerical methods,Dynamical low-rank approximation,Kinetic equation,Vlasov equation},
 }
 
-@software{tamelliniLorenzotamelliniSparsegridsmatlabkit2024,
-  title = {Lorenzo-Tamellini/Sparse-Grids-Matlab-Kit},
-  author = {family=tamellini, prefix=lorenzo-, useprefix=true},
-  date = {2024-12-05T13:17:27Z},
-  origdate = {2023-05-06T12:13:51Z},
-  url = {https://github.com/lorenzo-tamellini/sparse-grids-matlab-kit},
-  urldate = {2024-12-10},
-  abstract = {repository for the Sparse Grids Matlab Kit source code. Full info on https://sites.google.com/view/sparse-grids-kit}
+@software{TheSparseGridsMatlabKit,
+  title = {The Sparse Grids Matlab Kit},
+  author = {Tamellini, Lorenzo and Piazzola, Chiara and Nobile, Fabio and Sprungk, Björn and Porta, Giovanni and Guignard, Diane and Tesei, Francesco},
+  date = {2024-12-19},
+  url = {https://sites.google.com/view/sparse-grids-kit/home},
+  urldate = {2024-12-19},
+  abstract = {The Sparse Grids Matlab Kit provides a Matlab implementation of sparse grids, and can be used for approximating high-dimensional functions and, in particular, for surrogate-model-based uncertainty quantification.
+It is lightweight, high-level and (hopefully) easy to use, good for quick prototyping and teaching. It comes with a very extensive documentation and examples (8800 lines of code, 4800 lines of comments).}
 }
 
 @software{DisCoTeccombischemeutilities2022,
