### 2024-12-19 bab139a
paper: try @misc for software references according to https://joss.readthedocs.io/en/latest/example_paper.html
