+++ paper.bib
@@ -144,14 +144,15 @@
 It is lightweight, high-level and (hopefully) easy to use, good for quick prototyping and teaching. It comes with a very extensive documentation and examples (8800 lines of code, 4800 lines of comments).}
 }
 
-@software{DisCoTeccombischemeutilities2022,
+@misc{DisCoTec-combischeme-utilities,
+  author = {SG++ development team},
   title = {{{DisCoTec-combischeme-utilities}}},
-  date = {2022-10-18T05:44:12Z},
-  origdate = {2022-10-07T08:29:19Z},
-  url = {https://github.com/SGpp/DisCoTec-combischeme-utilities},
+  year = {2022},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/SGpp/DisCoTec-combischeme-utilities}, 
   urldate = {2023-03-21},
   abstract = {Python utilities to create (combination technique schemes) input files for the DisCoTec code},
-  organization = {SG++ development team}
 }
 
 @article{verboncoeurParticleSimulationPlasmas2005,
@@ -168,23 +169,26 @@ It is lightweight, high-level and (hopefully) easy to use, good for quick protot
   urldate = {2024-12-04},
   issue = {5A},
 }
-@software{SGppSGpp2024,
-  title = {{{SGpp}}/{{SGpp}}},
-  date = {2024-12-05T10:19:33Z},
-  origdate = {2018-03-05T12:33:52Z},
+
+@misc{SGpp,
+  title = {{SGpp}},
+  author = {SG++ development team},
+  title = {Fidgit: An ungodly union of GitHub and Figshare},
+  year = {2018},
+  publisher = {GitHub},
+  journal = {GitHub repository},
   url = {https://github.com/SGpp/SGpp},
   urldate = {2024-12-10},
-  abstract = {SG⁺⁺ – the numerical library for Sparse Grids in all their variants.},
-  organization = {SG++ development team},
-  keywords = {adaptive-sparse-grids,b-splines,cpp,interpolation,java,machine-learning,matlab,numerics,optimization,pde,python,quadrature,regression,sparse-grids}
+  abstract = {SG⁺⁺ -- the numerical library for Sparse Grids in all their variants.}
 }
 
-@software{obersteinerSparseSpACESparseGrid2023,
-  title = {{{sparseSpACE}} - {{The Sparse Grid Spatially Adaptive Combination Environment}}},
+@misc{sparseSpACE,
   author = {Obersteiner, Michael},
-  date = {2023-05-09T06:55:23Z},
-  origdate = {2019-09-02T09:22:49Z},
-  url = {https://github.com/obersteiner/sparseSpACE},
+  title = {{{sparseSpACE}} - {{The Sparse Grid Spatially Adaptive Combination Environment}}},
+  year = {2019},
   urldate = {2023-08-07},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/obersteiner/sparseSpACE},
   abstract = {sparseSpACE - the Sparse Grid Spatially Adaptive Combination Environment implements different variants of the spatially adaptive combination technique.}
 }
