### 2024-12-19 843db75
paper: fix bib title capitalization
