+++ paper.bib
@@ -135,7 +135,7 @@
 }
 
 @software{TheSparseGridsMatlabKit,
-  title = {The Sparse Grids Matlab Kit},
+  title = {{The Sparse Grids Matlab Kit}},
   author = {Tamellini, Lorenzo and Piazzola, Chiara and Nobile, Fabio and Sprungk, Björn and Porta, Giovanni and Guignard, Diane and Tesei, Francesco},
   date = {2024-12-19},
   url = {https://sites.google.com/view/sparse-grids-kit/home},
@@ -173,7 +173,6 @@ It is lightweight, high-level and (hopefully) easy to use, good for quick protot
 @misc{SGpp,
   title = {{SGpp}},
   author = {SG++ development team},
-  title = {Fidgit: An ungodly union of GitHub and Figshare},
   year = {2018},
   publisher = {GitHub},
   journal = {GitHub repository},
