### 2024-12-19 SGpp/DisCoTec#137 @freifrauvonbleifrei
good point about the references, I found JOSS' suggestions on it: https://joss.readthedocs.io/en/latest/example_paper.html (though I have to admit that "In GitHub repository" reads a bit funny ;) )

updated paper:
https://github.com/SGpp/DisCoTec/actions/runs/12409073785/artifacts/2342042020
