### 2024-12-19 ede7dd8
docs: minor edits cf. https://github.com/SGpp/DisCoTec/issues/129#issuecomment-2545779773
