### 2024-12-19 9d03c7d
docs: notation fixes cf. https://github.com/SGpp/DisCoTec/issues/134#issuecomment-2547321135
