### 2024-12-19 SGpp/DisCoTec#134 @freifrauvonbleifrei
Ha, my markdown editor is not rendering some of the math at all. sorry about that!

should be fixed with the most recent commit.
