### 2025-01-06 SGpp/DisCoTec#140 @EmilyBourne
[JOSS] Compiling documentation
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

The documentation at https://discotec.readthedocs.io/ doesn't seem to include any information about how to compile.

In https://discotec.readthedocs.io/en/latest/simple_tutorial.html you say:
> Compile the code using your familiar build tools. MPI and DisCoTec should be the only immediate dependencies.

But this doesn't explain how to make DisCoTec into a dependency.

For a compilation with Makefile I imagine that we need to include the `src/` folder and `build-linux-ubuntu22.04-skylake-ak5g3b5/spack-build-ak5g3b5/src/libdiscotec.a`. Is this everything that is needed?

For a compilation with CMake, I am not sure how to proceed. Is there a way to link the package e.g. `find_package(DisCoTec)`? Or should I use `add_subdirectory(DisCoTec)`? Or something else
