### 2025-01-13 SGpp/DisCoTec#141 @EmilyBourne
[JOSS] Documentation needs updating in selalib_distributed/README.md
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

As discussed here:
https://github.com/SGpp/DisCoTec/pull/133#discussion_r1810750135
It should be possible to use SeLaLib as an example using the public version of SeLaLib. However the documentation [here](https://github.com/SGpp/DisCoTec/blob/main/examples/selalib_distributed/README.md) still says:

> public part available here https://github.com/selalib/selalib
> "our" currently used branch is https://gitlab.mpcdf.mpg.de/clapp/selalib/-/tree/exahd_communicator
> ...
> clone and check out right version (see above)

Reading this I understand the "right version" to be the currently used branch on gitlab.mpcdf.mpg.de this should be clarified
