### 2025-01-14 SGpp/DisCoTec#140 @freifrauvonbleifrei
excellent question! for the examples as a standalone target, we use `add_subdirectory` (cf. https://github.com/SGpp/DisCoTec/blob/9b5f190047decb848c3bb2613468b58bc2ef4941/examples/distributed_third_level/CMakeLists.txt#L15 ), but I suppose you are looking for an install step for DisCoTec? Let me think about how to do this!
