### 2025-01-15 SGpp/DisCoTec#140 @freifrauvonbleifrei
Ah, as per your other question @EmilyBourne : yes, only `libdiscotec.a` and the headers (and dependencies etc.) should be required.
