### 2025-01-16 43345c4
selalib example: update README (cf. https://github.com/SGpp/DisCoTec/issues/141 )
