+++ examples/selalib_distributed/README.md
@@ -1,10 +1,13 @@
-# Selalib Example
+# SeLaLib Example
 
 * runs the Semi-Lagrangian BSL code
-* public part available here <https://github.com/selalib/selalib>
-* "our" currently used branch is <https://gitlab.mpcdf.mpg.de/clapp/selalib/-/tree/exahd_communicator>
+* the example can be used with <https://github.com/selalib/selalib>;
+  it uses the SeLaLib target `sll_m_sim_bsl_vp_3d3v_cart_dd_slim_interface`
+* the experiments on mass conservation presented
+  [here](https://doi.org/10.1016/j.jcp.2023.112338)
+  used the branch <https://gitlab.mpcdf.mpg.de/clapp/selalib/-/tree/exahd_communicator> .
 
-## Build Selalib Dependency
+## Build SeLaLib Dependency
 
 * clone and check out right version (see above)
 * depends on `fftw` and `hdf5+fortran@1.10.5` or higher (in spack notation)
@@ -12,7 +15,7 @@
 * in the build directory, run
   
 ```sh
-cmake -DCMAKE_BUILD_TYPE=Release -DOPENMP_ENABLED=ON -DHDF5_PARALLEL_ENABLED=ON  -DUSE_FMEMPOOOL=OFF -DCMAKE_INSTALL_PREFIX=$(pwd)/install $SELALIB_DIR
+cmake -DCMAKE_BUILD_TYPE=Release -DOPENMP_ENABLED=ON -DHDF5_PARALLEL_ENABLED=ON -DUSE_FMEMPOOOL=OFF -DCMAKE_INSTALL_PREFIX=$(pwd)/install $SELALIB_DIR
 make test_cpp_interface
 make sll_m_sim_bsl_vp_3d3v_cart_dd_slim_interface
 make sim_bsl_vp_3d3v_cart_dd_slim
@@ -25,7 +28,7 @@ where `test_cpp_interface` can be used to test the gerenal C/Fortran interface,
 this `selalib_distributed` example.
 (This may take some time and usually fails if tried in parallel with `make -j`).
 
-### Build Selalib+DisCoTec example
+### Build SeLaLib+DisCoTec example
 
 * run DisCoTec's cmake with `-DDISCOTEC_WITH_SELALIB=1 -DSELALIB_DIR=` flags
   (`SELALIB_DIR` should be set to selalib's `CMAKE_INSTALL_PREFIX` folder, appended
@@ -33,7 +36,7 @@ this `selalib_distributed` example.
   `libselalib.a` and `libsll_m_sim_bsl_vp_3d3v_cart_dd_slim_interface.a`)
 * then `make selalib_distributed`
 
-### Run Selalib+DisCoTec example
+### Run SeLaLib+DisCoTec example
 
 * update combination technique parameters in `ctparam`, selalib parameters in
   `template/param.nml` (be careful with `haveDiagnosticsTask` -- it interpolates
@@ -42,7 +45,7 @@ this `selalib_distributed` example.
   MPI tasks has to match the process group number and size in `ctparam`.
   Make sure that OpenMP runs with 1 process per MPI rank.
 
-### Postprocessing for Selalib+DisCoTec example
+### Postprocessing for SeLaLib+DisCoTec example
 
 * run `./combine_selalib_diagnostics ctparam` or update paths in
   `./plot_landau.py` and run (which also invokes`./combine_selalib_diagnostics`
