### 2025-01-16 SGpp/DisCoTec#141 @freifrauvonbleifrei
@EmilyBourne I think this sounds more like what should be said here, do you agree?
