### 2025-01-16 SGpp/DisCoTec#141 @EmilyBourne
I agree. This seems clearer :+1:
