### 2025-01-16 SGpp/DisCoTec#140 @freifrauvonbleifrei
I just realized that @vancraar has already put a neat installation step here:
https://github.com/SGpp/DisCoTec/blob/9b5f190047decb848c3bb2613468b58bc2ef4941/src/CMakeLists.txt#L248
however, it is currently not installing the header files, only the library and all other executables.

So you can either go ahead with `add_subdirectory` or give us a few days to make `find_package` work @EmilyBourne
