### 2025-01-16 SGpp/DisCoTec#142 @EmilyBourne
[JOSS] Compilation problems
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

I am trying to compile the SeLaLib examples using the instructions found here:
https://github.com/SGpp/DisCoTec/commit/43345c4ad120b81ed4d094484606c0c6d87df453

I have successfully compiled and installed SeLaLib and have successfully run the `cmake` command to configure DisCoTec. However I am getting the following error:

```
In file included from /home/emily/Code/DisCoTec/src/fault_tolerance/LPOptimizationInterpolation.cpp:1:
/home/emily/Code/DisCoTec/src/fault_tolerance/LPOptimizationInterpolation.hpp:8:10: fatal error: glpk.h: No such file or directory
    8 | #include "glpk.h"
      |          ^~~~~~~~
compilation terminated.
```
I cannot see this `glpk.h` file in the DisCoTec folder. Maybe the CMake configuration is missing a dependency?
