### 2025-01-16 SGpp/DisCoTec#142 @EmilyBourne
It seems that GLPK is correctly requested here:
https://github.com/SGpp/DisCoTec/blob/main/src/CMakeLists.txt#L169

It is not clear why CMake did not raise an error when I had not loaded this module via spack
