### 2025-01-16 SGpp/DisCoTec#141 @EmilyBourne
I don't know if this is the right place for this or if you'd rather I created a dedicated issue.

I am having some trouble running the SeLaLib example. I have worked out that the example requires 9 MPI threads if it is to run without modification (8 + 1 world proc).
This is not clear from the docs:
> the number of MPI tasks has to match the process group number and size in ctparam

However if I copy the `param.nml` from `templates/param.nml` to my folder and execute I get the following errors:
```
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
At line 349 of file /home/emily/Code/selalib/simulations/parallel/bsl_vp_3d3v_cart_dd/sll_m_sim_bsl_vp_3d3v_cart_dd_slim.F90
Fortran runtime error: Cannot match namelist object name vp_b2_3d3v
```
Perhaps this is something that I need to fix from the SeLaLib side? Or that I need to define in the `param.nml` file? But it would be nice to have an example that works out of the box
