### 2025-01-16 SGpp/DisCoTec#144 @EmilyBourne
[JOSS] Can't run examples
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7018.

I have tried to run most of the examples in the folder https://github.com/SGpp/DisCoTec/tree/main/examples but I haven't had much success. I don't know if I have missed something important in the docs or if the READMEs for these files need updating.

## combi_example

No `README.md` provided.

Tried with
```
mpirun -n 5 ./combi_example
```
as suggested in `run.sh`

Error:
```
terminate called after throwing an instance of 'std::runtime_error'
terminate called after throwing an instance of 'std::runtime_error'
undup
  what():  Currently testing to not duplicate communicator (to save memory),                           if you do want to use this code please take care that                           MPI_Comm_free(&communicator_) will be called at some point
  what():  Currently testing to not duplicate communicator (to save memory),                           if you do want to use this code please take care that                           MPI_Comm_free(&communicator_) will be called at some point
```

## combi_example_faults

No `README.md` provided.

Tried with
```
mpirun -n 5 ./combi_example_faults
```
as suggested in `run.sh`

Error:
```
group 0combi_example_faults: /home/emily/Code/DisCoTec/src/utils/Stats.hpp:133: static void combigrid::Stats::setAttribute(const string&, const string&): Assertion `(initialized_)&&(str.str().c_str())' failed.
group 1combi_example_faults: /home/emily/Code/DisCoTec/src/utils/Stats.hpp:133: static void combigrid::Stats::setAttribute(const string&, const string&): Assertion `(initialized_)&&(str.str().c_str())' failed.
[emily-ThinkPad-T14-Gen-3:66340] *** Process received signal ***
[emily-ThinkPad-T14-Gen-3:66340] Signal: Aborted (6)
[emily-ThinkPad-T14-Gen-3:66340] Signal code:  (-6)
[emily-ThinkPad-T14-Gen-3:66340] [ 0] group 0combi_example_faults: /home/emily/Code/DisCoTec/src/utils/Stats.hpp:133: static void combigrid::Stats::setAttribute(const string&, const string&): Assertion `(initialized_)&&(str.str().c_str())' failed.
[emily-ThinkPad-T14-Gen-3:66339] *** Process received signal ***
[emily-ThinkPad-T14-Gen-3:66339] Signal: Aborted (6)
[emily-ThinkPad-T14-Gen-3:66339] Signal code:  (-6)
[emily-ThinkPad-T14-Gen-3:66339] [ 0] group 1combi_example_faults: /home/emily/Code/DisCoTec/src/utils/Stats.hpp:133: static void combigrid::Stats::setAttribute(const string&, const string&): Assertion `(initialized_)&&(str.str().c_str())' failed.
[emily-ThinkPad-T14-Gen-3:66341] *** Process received signal ***
[emily-ThinkPad-T14-Gen-3:66341] Signal: Aborted (6)
[emily-ThinkPad-T14-Gen-3:66341] Signal code:  (-6)
[emily-ThinkPad-T14-Gen-3:66341] [ 0] /lib/x86_64-linux-gnu/libc.so.6(+0x42520)[0x73cf5aa42520]
group 2combi_example_faults: /home/emily/Code/DisCoTec/src/utils/Stats.hpp:133: static void combigrid::Stats::setAttribute(const string&, const string&): Assertion `(initialized_)&&(str.str().c_str())' failed.
```

## distributed_advection

No `README.md` or `run.sh` provided.

Tried with
```
mpirun -n 3 ./distributed_combi_example
```
as indicated by the error message that arises after trying with `mpirun -n 9 ./distributed_combi_example` as suggested by `template/submit.sh`

Error:
```
set up component grids and run until first combination point
terminate called after throwing an instance of 'std::runtime_error'
  what():  CFL condition not satisfied!
```

## distributed_third_level

No `README.md` provided.

Not sure how to test here as none of the generated files are called `combi_example` (which is the exe used in `run.sh`)
