### 2025-01-16 015aca8
Use a fatal error if FindGLPK is required but not found (#143)
