+++ CMakeModules/FindGLPK.cmake
@@ -25,6 +25,6 @@ if(GLPK_FOUND)
   endif(NOT GLPK_FIND_QUIETLY)
 else(GLPK_FOUND)
   if(GLPK_FIND_REQUIRED)
-    message(STATUS "FindGLPK: Could not find glpk.h and/or libglpk.a")
+    message(FATAL_ERROR "FindGLPK: Could not find glpk.h and/or libglpk.a")
   endif(GLPK_FIND_REQUIRED)
 endif(GLPK_FOUND)
