### 2025-01-20 c99e9fe
combi_example: fix, add readme cf. https://github.com/SGpp/DisCoTec/issues/144
