### 2025-01-20 #7018 @EmilyBourne
@danielskatz Just to give an update. I started trying to create my own test but then decided to start with the examples. Unfortunately I wasn't able to get any of them working so I am waiting for someone to get back to me to give me a hand in [[JOSS] Can't run examples SGpp/DisCoTec#144](https://github.com/SGpp/DisCoTec/issues/144)
