### 2025-01-21 b6c2bb3
advection_example: fix, add readme cf. https://github.com/SGpp/DisCoTec/issues/144
