### 2025-01-22 ec9a7ad
selalib example: fix, clean diagnostics script cf. https://github.com/SGpp/DisCoTec/issues/144 and https://github.com/SGpp/DisCoTec/issues/141
