### 2025-01-23 SGpp/DisCoTec#141 @freifrauvonbleifrei
@EmilyBourne I thought I had fixed the "namelist" error at some point, but could be wrong. It works in my current install with the github-selalib according to the updated README on branch `fix_examples`, would you mind trying again?
