### 2025-01-24 SGpp/DisCoTec#141 @EmilyBourne
@freifrauvonbleifrei I confirm that the selalib example now seems to be working. I will go back to looking through the examples in more detail and let you know if I come across any more problems
