### 2025-01-27 7bb9c5b
examples: some refinement
