### 2025-01-27 6f6dede
combi_workers_only example: add readme cf. https://github.com/SGpp/DisCoTec/issues/144
