### 2025-01-27 fad49db
third level example: fix, add readme cf. https://github.com/SGpp/DisCoTec/issues/144
