### 2025-01-27 SGpp/DisCoTec#144 @freifrauvonbleifrei
Hi @EmilyBourne , I think now most examples should be updated in branch `fix_examples`: if you like you can try `combi_example`, `combi_workers_only`, `distributed_advection`, `selalib_distributed` and `distributed_third_level` according to their respective READMEs.

apparently our CI was broken, so please give use a few more days to merge!
