### 2025-01-27 c1b687e
faults example: fix, initial readme cf. https://github.com/SGpp/DisCoTec/issues/144
