+++ examples/combi_example_faults/README.md
@@ -1,0 +1,14 @@
+# Example for Fault Tolerance
+
+In addition to the parameters in the [basic combi example](../combi_example/README.md),
+you can set the following to control the behavior of the random faults:
+
+```ini
+TODO!
+```
+
+Using the default settings, run with
+
+```bash
+mpirun -n 5 ./combi_example_faults
+```
