### 2025-01-28 2d8aebd
Refine documentation (#139)
* docs: address first two comments of
https://github.com/SGpp/DisCoTec/issues/129#issuecomment-2383274769

* docs: first draft for issue #134

* docs: minor edits to combination_technique

* docs: expand process group discussion

* docs: use 6 dimensions everywhere in tutorial
(addresses 6th point raised by @jakelangham in issue #129)

* docs: some more urls
(addresses last of @jakelangham's points in issue #129 )

* docs: another url in advanced_topics

* docs: minor edits
cf. https://github.com/SGpp/DisCoTec/issues/129#issuecomment-2545779773

* docs: notation fix
[…]
