### 2025-01-30 SGpp/DisCoTec#144 @EmilyBourne
@danielskatz yes this has helped. I was waiting for @freifrauvonbleifrei to inform me of the merge.

I will try and finish up my review on the train on Monday. I have now managed to run the examples so I am close to approving
