### 2025-02-04 1732d37
docs: add instruction for cmake/add_subdirectory cf. https://github.com/SGpp/DisCoTec/issues/140
