### 2025-02-04 SGpp/DisCoTec#140 @freifrauvonbleifrei
@EmilyBourne I have added the add_subdirectory instruction for now, the proper find_package setup is quite fiddly (esp. to get the paths nice and expressive). Will include the find_package instructions with #145, but we can leave that out of the review for now. Let me know what you think!
