### 2025-02-04 SGpp/DisCoTec#134 @freifrauvonbleifrei
closing this for now after #139 is merged, feel free to reopen if there's more to discuss!
