### 2025-02-12 #7018 @freifrauvonbleifrei
@danielskatz sorry, it's still our turn to move! currently waiting to get the definitive parameter description for this README https://github.com/SGpp/DisCoTec/pull/146/files#diff-d1c22b90059f86a60b02cc0c0f4b1a3d622d914e4825c510893842c34f47d6afR3 from coauthor @obersteiner who is expert on the features in this example. This is the only thing still missing for the PR merge, which I think is what @EmilyBourne is waiting for.
