### 2025-02-13 7cc4ac3
Update README.md
Added new parameters and description (plus reference).
