### 2025-02-13 c7c3f79
fixed compiler flag
