### 2025-02-13 26dd65d
faults example: README minor edits+format
