### 2025-02-14 e6a877c
Examples: Fix & add READMEs (#146)
* combi_example: fix, add readme
* advection_example: fix, add readme
* selalib example: fix, clean diagnostics script
* combi_workers_only example: add readme
* third level example: fix, add readme
* faults example: fix, README.md

cf. https://github.com/SGpp/DisCoTec/issues/144
and https://github.com/SGpp/DisCoTec/issues/141

---------

Co-authored-by: Theresa Pollinger <theresa.pollinger@riken.jp>
Co-authored-by: Michael Obersteiner <oberstei@in.tum.de>
