### 2025-02-17 #7018 @freifrauvonbleifrei
@danielskatz @EmilyBourne -- here's the ball back to you 🏐 the PR is merged and the issues are addressed as far as we are concerned, but feel free to note in case you see more things to address.

I would already like to extend a big thanks to @EmilyBourne in particular who looked into every dusty corner of the code and tested it, so we could fix all the issues you get when a certain setup is only ever used by a single person (who then moves on elsewhere after PhD). Sorry for not doing a better prior "dusting" ourselves and thx thx thx for helping! It has made the code **A LOT** better.
