### 2025-02-17 #7018 @freifrauvonbleifrei
Ah, we just realized that the paper.md is missing a section on funding acknowledgements, currently it has only people acknowledgements. Will provide within 1-2 days!
