### 2025-02-17 #7018 @danielskatz
> Ah, we just realized that the paper.md is missing a section on funding acknowledgements, currently it has only people acknowledgements. Will provide within 1-2 days!

ok, thanks - I'll make sure not to publish this until then 😄
