### 2025-02-18 4974142
paper.md: add funding acknowledgements
