### 2025-02-18 0e3f460
paper.md: fix link formatting
