### 2025-02-18 dfa8474
paper.bib: try to fix author formatting
