+++ paper.bib
@@ -145,7 +145,7 @@ It is lightweight, high-level and (hopefully) easy to use, good for quick protot
 }
 
 @misc{DisCoTec-combischeme-utilities,
-  author = {SG++ development team},
+  author = {{SG++ development team}},
   title = {{{DisCoTec-combischeme-utilities}}},
   year = {2022},
   publisher = {GitHub},
@@ -172,7 +172,7 @@ It is lightweight, high-level and (hopefully) easy to use, good for quick protot
 
 @misc{SGpp,
   title = {{SGpp}},
-  author = {SG++ development team},
+  author = {{SG++ development team}},
   year = {2018},
   publisher = {GitHub},
   journal = {GitHub repository},
