### 2025-02-19 #7018 @freifrauvonbleifrei
Now it's there ->  https://github.com/SGpp/DisCoTec/actions/runs/13385893884/artifacts/2607409136
