### 2025-02-19 471155c
minor changes in paper.md
