### 2025-02-19 a285c62
minor changes in paper.bib
