+++ paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{bungartzSparseGrids2004,
-  title = {Sparse {{Grids}}},
+  title = {Sparse Grids},
   author = {Bungartz, Hans-Joachim and Griebel, Michael},
   date = {2004},
   journaltitle = {Acta Numerica},
@@ -38,7 +38,7 @@
 }
 
 @inproceedings{pollingerRealizingJointExtremeScale2024,
-  title = {Realizing {{Joint Extreme-Scale Simulations}} on {{Multiple Supercomputers}}—{{Two Superfacility Case Studies}}},
+  title = {Realizing Joint Extreme-Scale Simulations on Multiple Supercomputers—Two Superfacility Case Studies},
   author = {Pollinger, Theresa and Craen, Alexander Van and Offenhäuser, Philipp and Pflüger, Dirk},
   date = {2024-11-09},
   pages = {1568--1584},
@@ -98,7 +98,7 @@
 
 @article{piazzolaSparseGridsMatlab2024,
  author = {Piazzola, Chiara and Tamellini, Lorenzo},
-  title  = {{Algorithm 1040: The Sparse Grids Matlab Kit - a Matlab implementation of sparse grids for high-dimensional function approximation and uncertainty quantification}},
+  title  = {Algorithm 1040: The Sparse Grids {M}atlab Kit - a {M}atlab implementation of sparse grids for high-dimensional function approximation and uncertainty quantification},
  journal= {ACM Transactions on Mathematical Software},
  year   = {2024},
  volume = {50},
@@ -135,7 +135,7 @@
 }
 
 @software{TheSparseGridsMatlabKit,
-  title = {{The Sparse Grids Matlab Kit}},
+  title = {The Sparse Grids {M}atlab Kit},
   author = {Tamellini, Lorenzo and Piazzola, Chiara and Nobile, Fabio and Sprungk, Björn and Porta, Giovanni and Guignard, Diane and Tesei, Francesco},
   date = {2024-12-19},
   url = {https://sites.google.com/view/sparse-grids-kit/home},
@@ -183,7 +183,7 @@ It is lightweight, high-level and (hopefully) easy to use, good for quick protot
 
 @misc{sparseSpACE,
   author = {Obersteiner, Michael},
-  title = {{{sparseSpACE}} - {{The Sparse Grid Spatially Adaptive Combination Environment}}},
+  title = {{sparseSpACE} - The Sparse Grid Spatially Adaptive Combination Environment},
   year = {2019},
   urldate = {2023-08-07},
   publisher = {GitHub},
