### 2025-02-19 #7018 @danielskatz
👋 @freifrauvonbleifrei - I've suggested some minor changes in https://github.com/SGpp/DisCoTec/pull/148 - please merge this, or let me know what you disagree with, then we can move on to the next steps.
