### 2025-02-20 #7018 @danielskatz
👋 @freifrauvonbleifrei - I see one thing I missed.

Can you change "as demonstrated in [@pollingerRealizingJointExtremeScale2024]" in the second paragraph to "as demonstrated in @pollingerRealizingJointExtremeScale2024"

Then, could you:
