### 2025-02-21 3b8999b
paper.md: minor edit
cf. https://github.com/openjournals/joss-reviews/issues/7018#issuecomment-2670083461
