### 2025-02-25 #7018 @freifrauvonbleifrei
@danielskatz Here's the Zenodo version!
https://doi.org/10.5281/zenodo.14920617
