### 2025-02-25 #7018 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
